% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_ae_class}
\alias{add_ae_class}
\title{Add Adverse Events class}
\usage{
add_ae_class(tbl, class)
}
\arguments{
\item{tbl}{(\code{tibble}) Containing the data}

\item{class}{(\code{character}) Class of adverse events to be added as
an rtable row}
}
\description{
Add Adverse Events class
}
