% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_pdf}
\alias{as_pdf}
\title{Output decorated grob (\code{gTree}) objects as PDF}
\usage{
as_pdf(grobs, outpath, pagesize = "letter.landscape")
}
\arguments{
\item{grobs}{a grid grob (\code{gTree}) object, optionally \code{NULL} if only a grob with
the decoration should be shown.}

\item{outpath}{specify full path to output pdf to \code{BCE} or \code{BEE}}

\item{pagesize}{name of \code{pagesize} (print size) and orientation, accepted values include
\code{"a4.landscape"}, \code{"a4.portrait"}, \code{"letter.portrait"} and
\code{"letter.landscape"} (default)}
}
\value{
a pdf file
}
\description{
This is an utility function to output a decorated grob (\code{gTree}) object
}
\examples{
\dontrun{
library(ggplot2)
g <- with(iris, {
  list(
    ggplotGrob(qplot(Sepal.Length, Sepal.Width, col = Species)),
    ggplotGrob(qplot(Sepal.Length, Petal.Length, col = Species)),
    ggplotGrob(qplot(Sepal.Length, Petal.Width, col = Species))
  )
})

# output to pdf
g \%>\% as_pdf("~/example_aspdf1.pdf")
}
}
\seealso{
\code{\link[=grobs2pdf]{grobs2pdf()}}
}
\author{
Chendi Liao (liaoc10) \email{chendi.liao@roche.com}
}
