% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_events_term_id.R
\name{create_flag_vars}
\alias{create_flag_vars}
\title{default ae overview flags}
\usage{
create_flag_vars(
  df,
  fatal = AESDTH == "Y",
  serious = AESER == "Y",
  serious_withdrawl = AESER == "Y" & grepl("DRUG WITHDRAWN", AEACN),
  serious_modified = AESER == "Y" & grepl("DRUG (INTERRUPTED|INCREASED|REDUCED)", AEACN),
  serious_related = AESER == "Y" & AEREL == "Y",
  withdrawl = grepl("DRUG WITHDRAWN", AEACN),
  modified = grepl("DRUG (INTERRUPTED|INCREASED|REDUCED)", AEACN),
  related = AEREL == "Y",
  related_withdrawl = AEREL == "Y" & grepl("DRUG WITHDRAWN", AEACN),
  related_modified = AEREL == "Y" & grepl("DRUG (INTERRUPTED|INCREASED|REDUCED)", AEACN),
  grade3 = AETOXGR \%in\% c("3", "4", "5"),
  ...
)
}
\arguments{
\item{df}{data frame of ae. use default}

\item{fatal}{AE with fatal outcome derivation}

\item{serious}{Serious AE derivation.}

\item{serious_withdrawl}{Serious AE leading to withdrawal derivation}

\item{serious_modified}{Serious AE leading to dose modification derivation}

\item{serious_related}{Related Serious AE derivation}

\item{withdrawl}{AE leading to withdrawal derivation}

\item{modified}{AE leading to dose modification derivation}

\item{related}{Related AE derivation}

\item{related_withdrawl}{Related AE leading to withdrawal derivation}

\item{related_modified}{Related AE leading to dose modification derivation}

\item{grade3}{Grade 3-5 AE derivation}

\item{...}{named expressions used to generate categories}
}
\description{
default ae overview flags
}
\details{
in this function, all flags are expressions calls, for simpler usage.
}
\examples{
library(random.cdisc.data)
ADAE <- cadae
adae_labels <- var_labels(ADAE)
adae_labels <- c(adae_labels,
                "Serious AE related to A: Drug X",
                "Serious AE related to B: Placebo")

# add AEACN to ADAE
AEACN <- sample(c("DOSE NOT CHANGED",
                  "DOSE INTERRUPTED",
                  "DOSE INCREASED",
                  "DOSE REDUCED",
                  "DRUG WITHDRAWN"),
                nrow(ADAE),
                replace = TRUE)
# add AESDTH to ADAE
AESDTH <- sample(c("Y", "N"),
                 nrow(ADAE),
                 replace = TRUE)

ADAE <- ADAE \%>\%
             # additional flag variable
             mutate(AEREL1 = (AEREL == "Y" &
                              ADAE$ACTARM == "A: Drug X")) \%>\%
             mutate(AEREL2 = (AEREL == "Y" &
                              ADAE$ACTARM == "B: Placebo"))
var_labels(ADAE) <- adae_labels
create_flag_vars(ADAE)
create_flag_vars(ADAE, `AENSER` = AESER != "Y") # create other flags
create_flag_vars(ADAE, fatal = NULL) # remove not needed flags
}
