% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_patient_profile.R
\name{domain_check}
\alias{domain_check}
\title{a helper function for g_patient_profile to check whether the domain has data available}
\usage{
domain_check(domain, name)
}
\arguments{
\item{domain}{domain input from g_patient_profile}

\item{name}{domain names}
}
\description{
a helper function for g_patient_profile to check whether the domain has data available
}
