% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_events_term_id.R
\docType{data}
\name{.datatable.aware}
\alias{.datatable.aware}
\title{allow data.table in pacakge}
\format{An object of class \code{logical} of length 1.}
\usage{
.datatable.aware
}
\description{
allow data.table in pacakge
}
\keyword{datasets}
