% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_ae_sub.R
\name{g_ae_sub}
\alias{g_ae_sub}
\title{Adverse Event Category Plot}
\usage{
g_ae_sub(
  term,
  term_selected,
  id,
  arm,
  arm_sl,
  subgroups,
  subgroups_sl,
  trt = unique(arm)[1],
  ref = unique(arm)[2],
  indent = 4,
  subgroups_labels = NULL,
  xmax = 0,
  ci_method = "wald",
  conf_level = 0.95,
  fontsize = 4,
  draw = TRUE
)
}
\arguments{
\item{term}{\code{character} event term vector.}

\item{term_selected}{A \code{character} or \code{vector} include selected term(s) that is passed to analysis.
All terms are included by default.}

\item{id}{\code{vector} id of the event term vector.
For example, \code{ADAE$USUBJID}.}

\item{arm}{A \code{vector} of the treatment variable.
For example, \code{ADAE$ACTARM}.}

\item{arm_sl}{A \code{vector} of the subject level treatment variable.
For example, \code{ADSL$ACTARM}.}

\item{subgroups}{\code{data.frame} Variables to conduct analysis.}

\item{subgroups_sl}{\code{data.frame} Subject level variables to conduct analysis.
Usually from ADSL.}

\item{ci_method}{\code{character} The method used to calculate confidence interval.
Defalt is "wald".
Possible choices are methods supported in `DescTools::BinomDiffCI`.}

\item{conf_level}{\code{numeric} The confidence interval level, default set to 0.95.}

\item{fontsize}{a \code{numeric} value controls the size of the font.
Default is 4.}

\item{subgroups_levels}{A nested named \code{list} of variables to conduct analysis.
The names of the nested lists are used to show as the label.
The children lists should start with "Total" = variable label,
followed by labels for each level of said variable. See example for reference.}
}
\value{
grob object
}
\description{
Draw adverse event category plot.
}
\details{
there is no equivalent STREAM output
}
\examples{
library(osprey)
library(random.cdisc.data)

ADAE <- cadae
ADSL <- cadsl

term <- as.character(ADAE$AEDECOD)
id <- ADAE$USUBJID
arm <- as.character(ADAE$ACTARMCD)
arm_sl <- as.character(ADSL$ACTARMCD)
subgroups_sl <- ADSL[, c("SEX", "RACE", "STRATA1")]
subgroups <- ADAE[, c("SEX", "RACE", "STRATA1")]
trt <- "ARM A"
ref <- "ARM C"
subgroups_labels <- list(RACE = list("Total" = "Race",
                                     "AMERICAN INDIAN OR ALASKA NATIVE" = "American",
                                     "WHITE" = "White",
                                     "ASIAN" = "Asian",
                                     "African" = "BLACK OR AFRICAN AMERICAN"),
                         STRATA1 = list("Total" = "Strata",
                                        "A" = "TypeA",
                                        "B" = "TypeB",
                                        "C" = "Typec"),
                         SEX = list("Total" = "Sex",
                                    "M" = "Male",
                                    "F" = "Female",
                                    "U" = "Unknown"))
term_selected <- unique(term)[1]
g_ae_sub(term,
         term_selected,
         id,
         arm,
         arm_sl,
         subgroups,
         subgroups_sl,
         subgroups_labels = subgroups_labels)
}
\author{
Liming Li (Lil128) \email{liming.li@roche.com}
}
