% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_butterfly_modD.R
\name{g_butterfly_modD}
\alias{g_butterfly_modD}
\title{Create a butterfly plot for Early Development Visualization}
\usage{
g_butterfly_modD(category, groups, group_names = NULL,
  block_count = "# of patients", block_color = NULL, id,
  facet_rows = NULL, x_label = block_count, y_label = "AE Derived Terms",
  legend_label = "AETOXGR", sort_by = "alphabetical", show_legend = TRUE)
}
\arguments{
\item{category}{vector of y values}

\item{groups}{2 column dataframe  of 1/0 (flags of dichotomization values)}

\item{group_names}{string vector of length 2 with desired names of dichotomization variables
required format : first name corresponds to the name of the first column in group
                  second name corresponds to name of second column in group
default: will extract column names from group}

\item{block_count}{string - what to count by (ex: # of AEs or # of patients)}

\item{block_color}{vector - color coding of bar segments}

\item{id}{unique subject identifier variable.}

\item{facet_rows}{vector defines what variable is used to split the
plot into rows, default here is NULL}

\item{x_label}{string of text for x axis label, default is block_count}

\item{y_label}{string of text for y axis label, default is AE Derived Terms}

\item{sort_by}{character string that defines the ordering of the class and term
variables in the output table,
options: "alphabetical" or "count", default here is set to "count"}

\item{show_legend}{boolean of whether color coding legend is included,
default here is FALSE}
}
\value{
ggplot object
}
\description{
Create a butterfly plot for Early Development Visualization
}
\examples{
library(random.cdisc.data)
library(dplyr)

AAE <- radam("AAE", N=10)
ADSL <- radam("ADSL", N=10)
ANL <- left_join(AAE, ADSL, by = c("STUDYID", "USUBJID"))
ANL <- ANL \%>\% mutate(flag1 = ifelse(SEX == "F", 1, 0)) \%>\% mutate(flag2 = ifelse(ARM == "ARM A", 1, 0))

g_butterfly_modD(category = ANL$AEBODSYS,
            groups = data.frame(flag1 = ANL$flag1, flag2 = ANL$flag2),
            group_names = c("flag1", "flag2"),
            block_count = "# of patients",
            block_color = ANL$AETOXGR,
            id = ANL$USUBJID,
            facet_rows = ANL$RACE,
            x_label = "# of patients",
            y_label = "AE Derived Terms",
            legend_label = "AETOXGR",
            sort_by = "count",
            show_legend = TRUE)

}
\author{
Carolyn Zhang
}
