% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_events_term_id.R
\name{g_events_term_id}
\alias{g_events_term_id}
\title{Common Events Plot}
\usage{
g_events_term_id(
  term,
  id,
  arm,
  arm_sl,
  trt = unique(arm)[1],
  ref = unique(arm)[2],
  sort_by = "term",
  rate_range = c(0, 1),
  diff_range = c(-1, 1),
  reversed = FALSE,
  conf_level = 0.95,
  ci_method = "wald",
  axis_side = "left",
  color = c("red", "blue"),
  shape = c(16, 17),
  fontsize = 4,
  draw = TRUE
)
}
\arguments{
\item{term}{\code{character} or \code{factor} vector represents events information}

\item{id}{\code{vector} contains subject identifier. Length of \code{id} must be the same as
the length or number of rows of \code{terms}. Ususally it is \code{USUBJID}.}

\item{arm}{\code{character} vector contains arm informatiion. For example, \code{ACTARMCD}.}

\item{arm_sl}{\code{character} vector contains subject level arm vector. For example, \code{ADSL$ACTARMCD}.}

\item{trt}{\code{character} indicates the name of the treatment arm}

\item{ref}{\code{character} indicates the name of the reference arm}

\item{sort_by}{\code{character} indicates sort_by type. Choose from "term", "riskdiff"
and "meanrisk". Default is "term".}

\item{rate_range}{Range for overall rate}

\item{diff_range}{Range for rate difference}

\item{reversed}{Whether to sort_by on reversed variable. Default set to FALSE.}

\item{conf_level}{The confidence interval level, default set to 0.95.}

\item{ci_method}{The method used to calculate confidence interval. Defalt is "wald".
Possible choices are methods supported in `DescTools::BinomDiffCI`.}

\item{axis_side}{the side of the axis label, "left" or "right". Default set to "left".}

\item{color}{Color for the plot. Vector of length 2. Color for arms seperately.
Default set to c("red", "blue")}

\item{shape}{Shape for the plot. Vector of length 2. Shape for arms seperately.
Default set to c(16, 17).}

\item{fontsize}{font size for the plot. It is the size used in ggplot2 with default unit "mm", if you
want "points" you will need to devide the point number by "ggplot2:::.pt"}

\item{draw}{whether to draw the Plot.}
}
\value{
grob object
}
\description{
This function plots commonly occurred events by number of unique subjects with events.
It creates basic summary of events and compares event occurrences between treatment and reference
arms, and can be used for events data such as Adverse Events.
}
\details{
there is no equivalent STREAM output
}
\examples{
library(osprey)
library(random.cdisc.data)

term <- as.character(cadae$AEDECOD)
id <- cadae$USUBJID
arm <- as.character(cadae$ACTARMCD)
arm_sl <- as.character(cadsl$ACTARMCD)
trt <- "ARM A"
ref <- "ARM B"
g_events_term_id(term, id, arm, arm_sl, sort_by = "riskdiff")
g_events_term_id(term, id, arm, arm_sl, sort_by = "term", reversed = FALSE)
g_events_term_id(term, id, arm, arm_sl,
  sort_by = "meanrisk",
  reversed = FALSE, axis_side = "right"
)
g_events_term_id(term, id, arm, arm_sl, conf_level = 0.9, fontsize = 6)
term <- create_flag_vars(cadae)
g_events_term_id(term, id, arm, arm_sl, conf_level = 0.9, fontsize = 4)
}
\author{
Liming Li (Lil128) \email{liming.li@roche.com}

Molly He (hey59) \email{hey59@gene.com}
}
