% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_heat_bygrade.R
\name{g_heat_bygrade}
\alias{g_heat_bygrade}
\title{Heatmap by Grade}
\usage{
g_heat_bygrade(
  exp_data,
  anno_data,
  anno_var,
  heat_data,
  heat_color_var,
  heat_color_opt,
  conmed_data,
  conmed_var,
  xlab = "Visit",
  title = "Heatmap by Grade"
)
}
\arguments{
\item{exp_data}{(`dataframe`)\cr contains subject identifier. Usually it is \code{ADSL$USUBJID}.}

\item{anno_data}{(`dataframe`)\cr to specify color for the heatmap plot.
For example \code{ADAE$AETOXGR} or \code{ADCE$CETOXGR}.}

\item{anno_var}{(`vector`)\cr}

\item{heat_data}{(`dataframe`)\cr contains the information needed for the text over heatmap
Usually is \code{ADCM}.}

\item{heat_color_var}{(`character`)\cr}

\item{heat_color_opt}{(`vector`)\cr}

\item{conmed_data}{(`dataframe`)\cr}

\item{conmed_var}{(`character`)\cr}

\item{xlab}{(`character`)\cr}

\item{title}{(`character`)\cr}
}
\description{
This function plots heatmap
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 30)
ADEX <- radex(ADSL)
ADAE <- radae(ADSL)
ADCM <- radcm(ADSL)
# add AVISIT in ADAE
ADAE$AVISIT <- sample(unique(ADEX$AVISIT[!is.na(ADEX$AVISIT)]), nrow(ADAE), TRUE)
# include three conmed levels
ADCM_labs <- rtables::var_labels(ADCM)
ADCM <- ADCM \%>\%
  filter(
    CMDECOD == "medname A_1/3" | CMDECOD == "medname A_2/3" | CMDECOD == "medname A_3/3"
    ) \%>\%
  mutate(CMDECOD = factor(CMDECOD, levels = unique(CMDECOD)))
rtables::var_labels(ADCM) <- ADCM_labs
exp_data <- ADEX \%>\% filter(PARCAT1 == "INDIVIDUAL")
anno_data <- ADSL
anno_var <- c("SEX","COUNTRY")
heat_data <- ADAE
heat_color_var <- "AETOXGR"
heat_color_opt <- c("0" = "gray90","1" = "lightsteelblue1", "2" = "steelblue1", "3" = "steelblue4",
                    "4" = "maroon", "5" = "brown4")
conmed_data <- ADCM
conmed_var <- "CMDECOD"

xlab = "Visit"; title = "Heatmap by Grade"

g_heat_bygrade(
  exp_data,
  anno_data,
  anno_var,
  heat_data,
  heat_color_var,
  heat_color_opt,
  conmed_data,
  conmed_var
  )
}
\author{
Nina Qi (qit3) \email{qit3@gene.com}

Molly He (hey59) \email{hey59@gene.com}
}
