% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_waterfall.R
\name{g_waterfall}
\alias{g_waterfall}
\title{Waterfall Plot}
\usage{
g_waterfall(bar_id, bar_height, sort_by = NULL, col_by = NULL,
  bar_color_opt = NULL, anno_txt = NULL, href_line = NULL,
  facet_by = NULL, show_datavalue = TRUE, add_label = NULL,
  gap_point = NULL, ytick_at = 20,
  y_label = "Best \% Change from Baseline", title = "Waterfall Plot")
}
\arguments{
\item{bar_id}{vector of IDs to identify each bar}

\item{bar_height}{numeric vector to be plotted as height of each bar}

\item{sort_by}{vector to sort bars, default is \code{NULL}}

\item{col_by}{vector to color bars, default is \code{NULL}}

\item{bar_color_opt}{aesthetic values to map color values (named vector to map color values to each name).
If not \code{NULL}, please make sure this contains all posible values for \code{col_by} values,
otherwise default ggplot color will be assigned, please note that \code{NULL} needs to be specified
in this case}

\item{anno_txt}{dataframe of subject-level variables to be displayed as annotation below the waterfall plot,
default is \code{NULL}}

\item{href_line}{numeric vector to plot horizontal reference lines, default is \code{NULL}}

\item{facet_by}{a vector to facet plot and annotation table, default is \code{NULL}}

\item{show_datavalue}{boolean of whether value of bar height is shown, default is \code{TRUE}}

\item{add_label}{vector of one subject-level variable to be added to each bar except for bar_height,
default is \code{NULL}}

\item{gap_point}{singular numeric value for adding bar break when some bars are significantly higher than
others, default is \code{NULL}}

\item{ytick_at}{optional bar height axis interval, default is 20}

\item{y_label}{label for bar height axis, default is "Best \% Change from Baseline"}

\item{title}{string to be displayed as plot title, default is "Waterfall Plot"}
}
\value{
plot object
}
\description{
Waterfall plot is often used in Early Development (ED) to present each individual patient’s best
response to a particular drug based on a parameter.
}
\examples{
library(tidyr)

# Example 1
ADSL <- rADSL[1:15,]
ADRS <- rADRS \%>\%
  filter(USUBJID \%in\% ADSL$USUBJID)
ADTR <- rADTR\%>\%
  filter(USUBJID \%in\% ADSL$USUBJID) \%>\%
  select(USUBJID, PCHG) \%>\%
  group_by(USUBJID) \%>\%
  slice(which.min(PCHG))

TR_SL <- ADSL \%>\%
  inner_join(ADTR, "USUBJID")

SUB_ADRS <- ADRS \%>\%
  filter(PARAMCD == "BESRSPI" | PARAMCD == "OBJRSPI") \%>\%
  select(USUBJID, PARAMCD, AVALC, AVISIT, ADY) \%>\%
  spread(PARAMCD, AVALC)

ANL <- TR_SL \%>\%
  left_join(SUB_ADRS,"USUBJID")

anno_txt_vars <- c("TRTDURD", "BESRSPI", "OBJRSPI", "SEX", "BMK2")

g_waterfall(
  bar_height = ANL$PCHG,
  bar_id = sub(".*-", "", ANL$USUBJID),
  col_by = ANL$SEX,
  sort_by = ANL$ARM,
  # bar_color_opt = c("F" = "red", "M" = "green", "U" = "blue"),
  anno_txt = ANL[, anno_txt_vars],
  facet_by = NULL,
  href_line = c(-30, 20),
  add_label = ANL$BESRSPI,
  ytick_at = 20,
  gap_point = NULL,
  show_datavalue = TRUE,
  y_label = "Best \% Change frome Baseline",
  title = "Waterfall Plot"
)

# Example 2 facetting
anno_txt_vars <- c("BESRSPI", "OBJRSPI")

g_waterfall(
  bar_id = sub(".*-", "", ANL$USUBJID),
  bar_height = ANL$PCHG,
  sort_by = ANL$COUNTRY,
  col_by = ANL$SEX,
  bar_color_opt = c("F" = "tomato", "M" = "skyblue3", "U" = "darkgreen"),
  anno_txt = ANL[, anno_txt_vars],
  facet_by = ANL$STRATA2,
  href_line = c(-30, 20),
  add_label = ANL$BESRSPI,
  ytick_at = 20,
  gap_point = 260,
  y_label = "Best \% Change frome Baseline",
  title = "Waterfall Plot"
)

 # Example 3 extreme value
ANL$PCHG[3] <- 99
ANL$PCHG[5] <- 199
ANL$PCHG[7] <- 599
ANL$BESRSPI[3] <- "PD"
ANL$BESRSPI[5] <- "PD"
ANL$BESRSPI[7] <- "PD"

g_waterfall(
  bar_id = sub(".*-", "", ANL$USUBJID),
  bar_height = ANL$PCHG,
  sort_by = ANL$ARM,
  col_by = ANL$SEX,
  bar_color_opt = c("F" = "tomato", "M" = "skyblue3", "U" = "darkgreen"),
  anno_txt = ANL[, anno_txt_vars],
  facet_by = NULL,
  href_line = c(-30, 20),
  add_label = ANL$BESRSPI,
  ytick_at = 20,
  gap_point = 260,
  y_label = "Best \% Change frome Baseline",
  title = "Waterfall Plot"
)

}
\author{
Xuefeng Hou (houx14) \email{houx14@gene.com}

Ting Qi (qit3) \email{qit3@gene.com}
}
