% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_labels}
\alias{get_labels}
\title{Extracts dataset and variable labels from a dataset.}
\usage{
get_labels(data, fill = TRUE)
}
\arguments{
\item{data}{(`data.frame`) table to extract the labels from}

\item{fill}{(`logical(1)`) if `TRUE`, the function will return variable names for columns with non-existent labels;
otherwise will return `NA` for them}
}
\value{
column labels
}
\description{
Extracts dataset and variable labels from a dataset.
}
\keyword{internal}
