% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcefunctions.R
\name{getdata_bce}
\alias{getdata_bce}
\title{Load Multiple SAS datasets from BCE}
\usage{
getdata_bce(snapshot, area, x = NULL, load = TRUE)
}
\arguments{
\item{snapshot}{name of the BCE snapshot in string format (eg. "s12345a")}

\item{area}{specify the BCE area in string format, accepted values are
\code{"sdtm"}, \code{"prod"}, \code{"qa"} and \code{"home"}}

\item{x}{a character vector containing the name of the SAS datasets to be
loaded. Default is \code{NULL}, where all datasets within the snapshot
directory will be loaded.}

\item{load}{if \code{TRUE} (default), all read-in datasets will be loaded.
Use \code{FALSE} if do not want to load datasets automatically, and assign
to an object.}
}
\value{
dataframe objects
}
\description{
This is an utility function to quickly read-in one or multiple SAS datasets
from BCE without typing out the full path, while keeping the file name as the
assigned dataset name.
}
\examples{

\dontrun{
# Example 1 - load add datasets in directory
getdata_bce("s39148b", "prod")

# Example 2 - load only specified datasets
getdata_bce("s39148b", "sdtm", c("dm", "ex", "ds"))

# Example 3 - load from home
# equivalent to read_bce('/opt/BIOSTAT/home/<UNIXID>/<snapshot>/libraries/asl.sas7bdat')
snapshot <- "s39148b"
getdata_bce(snapshot, "home", "asl")
}

}
\author{
Chendi Liao (liaoc10) \email{chendi.liao@roche.com}
}
