% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grob_add_padding}
\alias{grob_add_padding}
\title{Add padding to grob}
\usage{
grob_add_padding(grob, pad_v = unit(5, "pt"), pad_h = unit(5, "pt"))

grob_add_padding(grob, pad_v = unit(5, "pt"), pad_h = unit(5, "pt"))
}
\arguments{
\item{grob}{grob object}

\item{pad_v}{padding to add vertically}

\item{pad_h}{padding to add horizontally}
}
\description{
Add padding to grob

Add padding to grob
}
