% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grob_part}
\alias{grob_part}
\title{Extract specific part of a \code{ggplot} or grob}
\usage{
grob_part(gplot_grob, part)
}
\arguments{
\item{gplot_grob}{\code{ggplot} or grob object}

\item{part}{name of the part to be extracted. \code{NA} will return \code{zeroGrob()}}
}
\description{
Extract specific part of a \code{ggplot} or grob
}
\keyword{internal}
