% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grob_parts}
\alias{grob_parts}
\title{Extract specific parts of a ggplot or grob}
\usage{
grob_parts(gplot, parts)
}
\arguments{
\item{gplot}{ggplot or grob object}

\item{part}{names vector of the parts to be extracted.}
}
\description{
Extract specific parts of a ggplot or grob
}
