% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcefunctions.R
\name{grobs2pdf}
\alias{grobs2pdf}
\title{Decorate grob (gTree) objects then outputs as IDM compatible PDF}
\usage{
grobs2pdf(grobs, titles, footnotes, progpath, outpath, fontsize = 9,
  pagesize = "letter.landscape")
}
\arguments{
\item{grobs}{A grid grob (gTree) object, optionally \code{NULL} if only a
grob with the decoration should be shown}

\item{titles}{Vector of character strings. Vector elements are separated by a
newline and strings are wrapped according to the page with}

\item{footnotes}{Vector of character string. Same rules as for \code{titles}}

\item{progpath}{Specify the full path to the R program that generate the
grobs and the PDF}

\item{outpath}{Specify full path to output pdf to BCE or BEE}

\item{fontsize}{Base font size used in pdf, default set to 9. Font size for
title is set to \code{fontsize} + 1 (default = 10) and for footnotes set to
\code{fontsize} - 1 (default = 8)}

\item{pagesize}{name of paper size and orientation, accepted values include
\code{"a4.landscape"}, \code{"a4.portrait"}, \code{"letter.portrait"} and
\code{"letter.landscape"} (default)}
}
\value{
a pdf file
}
\description{
This is an utility function to decorated grob (gTree) object with titles and
footnotes in accordance with IDM specification and export as PDF file with
full path to program and the output for easy tracking and archiving.
}
\examples{

\dontrun{
library(ggplot2)

g <- with(iris, {
  list(
    ggplotGrob(qplot(Sepal.Length, Sepal.Width, col = Species)),
    ggplotGrob(qplot(Sepal.Length, Petal.Length, col = Species)),
    ggplotGrob(qplot(Sepal.Length, Petal.Width, col = Species))
  )
})

grobs2pdf(
  grobs = g,
  titles = "Visualization of Iris Data",
  footnotes = "This is a footnote",
  progpath = "~/example_prog.R",
  outpath = "~/example_grobs2pdf.pdf"
)
}
}
\author{
Chendi Liao (liaoc10) \email{chendi.liao@roche.com}
}
