% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meddra}
\alias{meddra}
\title{Meddra data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35333 rows and 8 columns.
}
\source{
internal
}
\usage{
data(meddra)
}
\description{
Meddra data
}
\keyword{datasets}
\keyword{internal}
