% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rADAE}
\alias{rADAE}
\title{Random adverse events}
\format{
An object of class \code{data.frame} with 297 rows and 48 columns.
}
\source{
internal
}
\usage{
data(rADAE)
}
\description{
Random adverse events
}
\keyword{datasets}
\keyword{internal}
