% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rADRS}
\alias{rADRS}
\title{Random response}
\format{
An object of class \code{data.frame} with 300 rows and 34 columns.
}
\source{
internal
}
\usage{
data(rADRS)
}
\description{
Random response
}
\keyword{datasets}
