% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rADTR}
\alias{rADTR}
\title{Random XXXX @todo}
\format{An object of class \code{data.frame} with 250 rows and 44 columns.}
\usage{
data(rADTR)
}
\description{
Random XXXX @todo
}
\examples{
data(rADTR)
}
\keyword{datasets}
