% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rADTTE}
\alias{rADTTE}
\title{Random Time to Event Analysis Dataset}
\format{
An object of class \code{data.frame} with 150 rows and 33 columns.
}
\source{
internal
}
\usage{
data(rADTTE)
}
\description{
Random Time to Event Analysis Dataset
}
\keyword{datasets}
