% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rATE}
\alias{rATE}
\title{Random ATE}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3600 rows and 66 columns.
}
\source{
internal
}
\usage{
data(rATE)
}
\description{
Random ATE
}
\keyword{datasets}
\keyword{internal}
