% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderplot.R
\name{spiderplot}
\alias{spiderplot}
\title{Create a spiderplot}
\usage{
spiderplot(x, y, col)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{col}{color of points}
}
\value{
ggplot object
}
\description{
Descr of this plot
}
\examples{
library(random.cdisc.data)


ASL <- radam("ASL", start_with = list(WEIGHT = c(80, 85, 90), STUDYID = "A"))

with(ASL, spiderplot(WEIGHT, BAGE, ARM))

spiderplot(x = ASL$WEIGHT, y = ASL$BAGE, col = ASL$ARM)


}
\author{
Mika
}
