% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderplot_simple.R
\name{spiderplot_simple}
\alias{spiderplot_simple}
\title{Create a spiderplot_simple}
\usage{
spiderplot_simple(anl)
}
\arguments{
\item{anl}{Analysis dataset}
}
\value{
ggplot object
}
\description{
Descr of this plot
}
\examples{
library(random.cdisc.data)


atr <- radam("ATR", N=10 ,start_with = list(CHANGE = c(0, 80, 85, 90), STUDYID = c("A","B")))


spiderplot_simple(atr \%>\% filter(PARAMCD == "SUMTGLES"))


}
\author{
Mika
}
