% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_filter.R
\name{stream_filter}
\alias{stream_filter}
\title{Applies STREAM style filtering to datasets}
\usage{
stream_filter(
  slref = NULL,
  anl = NULL,
  filters,
  suffix,
  slref_keep = NULL,
  usubjid = "USUBJID"
)
}
\arguments{
\item{slref}{The subject level data frame (typically ASL.sas7bdat on BCE)}

\item{anl}{The analysis data frame (e.g. ATE.sas7bdat on BCE)}

\item{filters}{The name of the filters dataset (typically __filters.sas7bdat on BCE)}

\item{suffix}{The suffix to apply in quotes (e.g. "ITT_PFSINV")}

\item{slref_keep}{Variables to keep from slref (e.g. c("REGION", "SEX"))}

\item{usubjid}{The unique subject identifier variable in quotes (e.g. "USUBJID")}
}
\value{
\code{dataframe} object
}
\description{
One of slref or anl need to be specified. The conversion from SAS code in filters dataset may not work in all cases.
In case of failure a sensible error message should be returned.
}
\examples{
library(random.cdisc.data)
ASL <- radsl()
ATE <- radaette(ASL)
filters <- as.data.frame(rbind(
  c(ID = "IT", FLTTARGET = "SLREF", FLTWHERE = "where 1 eq 1"),
  c(ID = "BIO", FLTTARGET = "SLREF", FLTWHERE = "where BMRKR1 ge 4.3"),
  c(ID = "M", FLTTARGET = "SLREF", FLTWHERE = "where SEX eq 'M'"),
  c(ID = "PFS", FLTTARGET = "ANL", FLTWHERE = "where PARAMCD eq 'PFS'"),
  c(ID = "OS", FLTTARGET = "ANL", FLTWHERE = "where PARAMCD eq 'OS'")
))

ANL <- stream_filter(
  slref = ASL,
  anl = ATE,
  suffix = "IT_PFS_BIO",
  filters = filters
)
}
\author{
Iain Bennett
}
