% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_filter.R
\name{stream_filter_index}
\alias{stream_filter_index}
\title{Replicates the use of index function in sas for logic options}
\usage{
stream_filter_index(string1, string2)
}
\arguments{
\item{string1}{The string to search within - can be a vector}

\item{string2}{The string to search for - must have length 1}
}
\value{
\code{boolean} indicator
}
\description{
Assumuption is that use in filters is to only resolve true vs false
Primarily for use with stream_filter and releated stream_filter_convwhere functions
}
\examples{
AEACN <- c("DRUG MODIFIED", "DRUG STOPPED", "DOSE/DRUG MODIFIED")
stream_filter_index(AEACN, "DRUG MODIFIED")
}
\author{
Iain Bennett
}
