% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{t_ae}
\alias{t_ae}
\title{Create an Adverse Events (AET02) Table}
\usage{
t_ae(class, term, id, col_by, total = "All Patients", sort_by = "count",
  ...)
}
\arguments{
\item{class}{system organ class variable.}

\item{term}{preferred term variable.}

\item{id}{unique subject identifier variable. If a particular subject has no
adverse event then the subject \code{id} should be listed where
\code{class} and \code{term} should be set to missing (i.e. \code{NA}).}

\item{col_by}{group variable that will be used for a column header. \code{col_by}
has to be a factor and can not be missing.}

\item{total}{character string that will be used as a label for a column with
pooled total population, default here is "All Patients", if set to "NONE" then
the "All Patients" column is suppressed.}

\item{sort_by}{character string that defines the ordering of the class and term
variables in the output table,
options: "alphabetical" or "count", default here is set to "count"}
}
\value{
\code{rtable} object
}
\description{
\code{t_ae} returns adverse events according to STREAM format AET02
}
\examples{
# Simple example
library(tibble)
library(dplyr)
library(rtables)

ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = paste("ARM", LETTERS[rep(c(1,2), c(3,7))])
)


ae_lookup <- tribble(
~CLASS,         ~TERM,   ~GRADE,
"cl A",   "trm A_1/2",        1,
"cl A",   "trm A_2/2",        2,
"cl B",   "trm B_1/3",        2,
"cl B",   "trm B_2/3",        3,
"cl B",   "trm B_3/3",        1,
"cl C",   "trm C_1/1",        1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2,2,2,3,3,4,4,4,4,5,6,6,7,7)]
  ),
  ae_lookup[c(1,1,2,6,4,2,2,3,4,2,1,5,4,6),]
)

ANL <- left_join(ASL, AAE, by = "USUBJID")

tbl <- t_ae(
  class = ANL$CLASS,
  term =  ANL$TERM,
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  total = "All Patients",
)

tbl

# Simple example 2

library(dplyr)
library(rtables)
library(random.cdisc.data)


data <- left_join(radam("AAE", N=10),radam("ADSL", N=10))

tbl2 <- t_ae(
  class = data$AEBODSYS,
  term =  data$AEDECOD,
  id = data$USUBJID,
  col_by = factor(data$ARM),
  total = "None",
  sort_by = "alphabetical"
)

tbl2


}
\author{
Carolyn Zhang
}
