% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae_oview.R
\name{t_ae_oview}
\alias{t_ae_oview}
\title{Create an AE Overview Summary Table (AET01)}
\usage{
t_ae_oview(id, class, term, dthfl, dcsreas, aesdth, aeser, aeacn, arel, aerel,
  aetoxgr, col_by, total = "All Patients")
}
\arguments{
\item{id}{unique subject identifier variable. If a particular subject has no
adverse event then the subject \code{id} should be listed where
\code{class} and \code{term} should be set to missing (i.e. \code{NA}).}

\item{class}{system organ class variable.}

\item{term}{preferred term variable.}

\item{dthfl}{subject death flag variable.}

\item{dcsreas}{reason for discontinuation from study variable.}

\item{aesdth}{results in death variable.}

\item{aeser}{serious event variable.}

\item{aeacn}{action taken with study treatment variable.}

\item{arel}{analysis causality variable.}

\item{aerel}{causality variable.}

\item{col_by}{group variable that will be used for a column header. \code{col_by}
has to be a factor and can not be missing.}

\item{total}{character string that will be used as a label for a column with
pooled total population, default here is "All Patients", if set to "NONE" then
the "All Patients" column is suppressed.}

\item{atoxgr}{analysis toxicity grade variable.}
}
\value{
\code{rtable} object
}
\description{
\code{t_ae_oview} returns adverse events according to STREAM format AET01
}
\examples{
library(dplyr)
suppressPackageStartupMessages(library(tidyverse))
library(rtables)
library(haven)
#adae = read_sas("/opt/bee/home_nas/zhanc107/osprey/adam/adae.sas7bdat")
adae <- read_bce("/opt/BIOSTAT/home_ext2/qit3/cdt70194/go39733/libraries/adae.sas7bdat")

tbl <- t_ae_oview(
   id = adae$USUBJID,
   class = adae$AESOC,
   term = adae$AEDECOD,
   dthfl = adae$DTHFL,
   dcsreas = adae$DCSREAS,
   aesdth = adae$AESDTH,
   aeser = adae$AESER,
   aeacn = adae$AEACN,
   arel = adae$AREL,
   aerel = adae$AEREL,
   aetoxgr = adae$AETOXGR,
   col_by = factor(adae$ARM),
   total="All Patients"
)

tbl

tb2 <- t_ae_oview(
   id = adae$USUBJID,
   class = adae$AESOC,
   term = adae$AEDECOD,
   dthfl = adae$DTHFL,
   dcsreas = adae$DCSREAS,
   aesdth = adae$AESDTH,
   aeser = adae$AESER,
   aeacn = adae$AEACN,
   arel = adae$AREL,
   aerel = adae$AEREL,
   aetoxgr = adae$AETOXGR,
   col_by = factor(adae$ARM),
   total="NONE"
)

tb2


}
\author{
Carolyn Zhang
}
