% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ds.R
\name{t_ds}
\alias{t_ds}
\title{Create a Patient Disposition (DST01) Table}
\usage{
t_ds(id, term, col_by, total = "All Patients")
}
\arguments{
\item{id}{unique subject identifier variable. If a particular subject has no
adverse event then the subject \code{id} should be listed where
\code{class} and \code{term} should be set to missing (i.e. \code{NA}).}

\item{term}{reason for discontinuation from study.}

\item{col_by}{group variable that will be used for a column header. \code{col_by}
has to be a factor and can not be missing.}

\item{total}{character string that will be used as a label for a column with
pooled total population, default here is "All Patients", if set to "NONE" then
the "All Patients" column is suppressed.}
}
\value{
\code{rtable} object
}
\description{
\code{t_ae} returns patient disposition according to STREAM format DST01
}
\examples{
# Simple example

library(dplyr)
library(rtables)
library(random.cdisc.data)


adae <- read.bce("/opt/BIOSTAT/home/bundfuss/stream_um/str_para2/libraries/adae.sas7bdat")

tbl <- t_ds(id = adae$USUBJID,
            term = adae$DCSREAS,
            col_by = factor(adae$ARM),
            total = "All Patients")

tbl
}
\author{
Carolyn Zhang
}
