% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_osprey}
\alias{theme_osprey}
\title{this theme is used across many figures. can be safely removed if update the theme in each function}
\usage{
theme_osprey(axis_side = "left", fontsize = 4, blank = FALSE)

theme_osprey(axis_side = "left", fontsize = 4, blank = FALSE)
}
\arguments{
\item{axis_side}{axis position}

\item{fontsize}{font size in 'mm'}

\item{blank}{whether to have blank or background with grids and borders}
}
\description{
this theme is used across many figures. can be safely removed if update the theme in each function

this theme is used across many figures. can be safely removed if update the theme in each function
}
\keyword{internal}
