% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variable_labels<-}
\alias{variable_labels<-}
\title{Sets column labels of a `data.frame`.}
\usage{
variable_labels(x) <- value
}
\arguments{
\item{x}{(`data.frame`) object with columns}

\item{value}{(`charater`) labels}
}
\value{
`x`
}
\description{
Sets column labels of a `data.frame`.
}
\examples{
variable_labels(iris) <- colnames(iris)
}
\keyword{internal}
