% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{waterfall}
\alias{waterfall}
\title{Create A Waterfall plot}
\usage{
waterfall(x, y)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}
}
\value{
a ggplot object
}
\description{
A waterfall plot shows...
}
\examples{
waterfall(1:5, 5:1)
}
