% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_events_term_id.R
\name{create_flag_vars}
\alias{create_flag_vars}
\title{default ae overview flags}
\usage{
create_flag_vars(
  df,
  `AE with fatal outcome` = AESDTH == "Y",
  `Serious AE` = AESER == "Y",
  `Serious AE leading to withdrawal` = AESER == "Y" & grepl("DRUG WITHDRAWN", AEACN),
  `Serious AE leading to dose modification` = AESER == "Y" &
    grepl("DRUG (INTERRUPTED|INCREASED|REDUCED)", AEACN),
  `Related Serious AE` = AESER == "Y" & AEREL == "Y",
  `AE leading to withdrawal` = grepl("DRUG WITHDRAWN", AEACN),
  `AE leading to dose modification` = grepl("DRUG (INTERRUPTED|INCREASED|REDUCED)",
    AEACN),
  `Related AE` = AEREL == "Y",
  `Related AE leading to withdrawal` = AEREL == "Y" & grepl("DRUG WITHDRAWN", AEACN),
  `Related AE leading to dose modification` = AEREL == "Y" &
    grepl("DRUG (INTERRUPTED|INCREASED|REDUCED)", AEACN),
  `Grade 3-5 AE` = AETOXGR \%in\% c("3", "4", "5"),
  ...
)
}
\arguments{
\item{df}{data frame of ae. use default}

\item{...}{named expressions used to generate categories}
}
\description{
default ae overview flags
}
\details{
in this function, all flags are expressions calls, for simpler usage.
}
\examples{
library(random.cdisc.data)
create_flag_vars(cadae)
create_flag_vars(cadae, `AENSER` = AESER != "Y") # create other flags
create_flag_vars(cadae, `Serious AE` = NULL) # remove not needed flags
}
