% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_ae_sub.R
\name{g_ae_sub}
\alias{g_ae_sub}
\title{Adverse Event Category Plot}
\usage{
g_ae_sub(
  term,
  term_selected,
  id,
  arm,
  arm_sl,
  subgroups,
  subgroups_sl,
  trt = unique(arm)[1],
  ref = unique(arm)[2],
  indent = 4,
  subgroups_labels = NULL,
  xmax = 0,
  cimethod = "wald",
  conf_level = 0.95,
  fontsize = 4,
  draw = FALSE
)
}
\arguments{
\item{term}{event term vector.}

\item{term_selected}{The selected term that is passed to analysis. Default set to the first value in term vector.}

\item{id}{id of the event term vector. Ususally it is USUBJID.}

\item{arm}{Treatment variable. Usually it is ACTARM}

\item{arm_sl}{Subject level treatment variable. Usually it is ACTARM}

\item{subgroups}{Variables to conduct analysis.}

\item{subgroups_sl}{Subject level variables to conduct analysis.}

\item{cimethod}{The method used to calculate confidence interval. Defalt is "wald".
Possible choices are methods supported in `DescTools::BinomDiffCI`.}

\item{conf_level}{The confidence interval level, default set to 0.95.}

\item{fontsize}{The size of the font. It is not the true font size but a multiplier.}

\item{col_by_levels}{Selected levels treatment and baseline respectively.}

\item{subgroups_levels}{Named list of variables to conduct analysis.
The names of the nested lists are used to show as the label.}

\item{title}{The title to be shown in the graph.}

\item{foot}{The footnote to be shown in the graph.}
}
\value{
grob object
}
\description{
Draw adverse event category plot.
}
\details{
there is no equivalent STREAM output
}
\examples{
library(osprey)
library(random.cdisc.data)

term <- as.character(cadae$AEDECOD)
id <- cadae$USUBJID
arm <- as.character(cadae$ACTARMCD)
arm_sl <- as.character(cadsl$ACTARMCD)
subgroups_sl <- cadsl[, c("SEX", "RACE", "STRATA1")]
subgroups <- cadae[, c("SEX", "RACE", "STRATA1")]
trt <- "ARM A"
ref <- "ARM C"
subgroups_labels <- list(RACE = list("Total" = "Race", "AMERICAN INDIAN OR ALASKA NATIVE" = "American", "WHITE" = "White", "ASIAN" = "Asian",
"African" = "BLACK OR AFRICAN AMERICAN"), STRATA1 = list("Total" = "Strata", "A" = "TypeA", "B" = "TypeB", "C" = "Typec"),
SEX = list("Total" = "Sex", "M" = "Male", "F" = "Female", "U" = "Unknown"))
term_selected <- unique(term)[1]
g_ae_sub(term, term_selected, id, arm, arm_sl, subgroups, subgroups_sl, subgroups_labels = subgroups_labels)
}
\author{
Liming Li (Lil128) \email{liming.li@roche.com}
}
