% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_ae_sub.R
\name{g_ae_sub}
\alias{g_ae_sub}
\title{Adverse Event Category Plot}
\usage{
g_ae_sub(
  term,
  id,
  arm,
  arm_sl,
  subgroups,
  subgroups_sl,
  trt = levels(arm)[1],
  ref = levels(arm)[2],
  indent = 4,
  subgroups_levels = NULL,
  xmax = 0,
  conf_level = 0.95,
  diff_ci_method = c("wald", "waldcc", "ac", "score", "scorecc", "mn", "mee", "blj",
    "ha", "beal"),
  fontsize = 4,
  draw = TRUE
)
}
\arguments{
\item{term}{\code{character} event term vector.}

\item{id}{(`vector`)\cr contains subject identifier. Length of \code{id} must be the
same as the length or number of rows of \code{terms}. Usually it is \code{ADAE$USUBJID}.}

\item{arm}{(`factor`)\cr vector that contains arm information in analysis data.
For example, `ADAE$ACTARMCD`.}

\item{arm_sl}{A \code{vector} of the subject level treatment variable.
For example, \code{ADSL$ACTARM}.}

\item{subgroups}{\code{data.frame} Variables to conduct analysis.}

\item{subgroups_sl}{\code{data.frame} Subject level variables to conduct analysis.
Usually from ADSL.}

\item{trt}{\code{character} indicates the name of the treatment arm. Default is the second
level of \code{arm}.}

\item{ref}{\code{character} indicates the name of the reference arm. Default is the first
level of \code{arm}.}

\item{indent}{\code{numeric} non-negative integer where 0 means that the subgroup levels should not be indented}

\item{subgroups_levels}{A nested named \code{list} of variables to conduct analysis.
The names of the nested lists are used to show as the label.
The children lists should start with "Total" = variable label,
followed by labels for each level of said variable. See example for reference.}

\item{xmax}{\code{numeric} maximum range for the x-axis.
x-axis range will be automatically assigned based on risk output when xmax is less than or equal to 0.
xmax is 0 by default}

\item{conf_level}{(`numeric`)\cr the confidence interval level, default is 0.95.}

\item{diff_ci_method}{(`character`)\cr the method used to calculate confidence interval.
Default is "wald". Possible choices are methods supported in \code{\link[DescTools]{BinomDiffCI}}.}

\item{fontsize}{(`numeric`)\cr font size for the plot. It is the size used in ggplot2 with
default unit "mm", if you want "points" you will need to divide the point number by
\code{ggplot2:::.pt}.}

\item{draw}{(`logical`)\cr whether to draw the plot.}
}
\value{
grob object
}
\description{
Draw adverse event category plot.
}
\details{
there is no equivalent STREAM output
}
\examples{
library(osprey)
library(random.cdisc.data)

ADAE <- radae(cached = TRUE)
ADSL <- radsl(cached = TRUE)

term <- as.character(ADAE$AEDECOD)
id <- ADAE$USUBJID
arm <- ADAE$ACTARMCD
arm_sl <- as.character(ADSL$ACTARMCD)
subgroups_sl <- ADSL[, c("SEX", "RACE", "STRATA1")]
subgroups <- ADAE[, c("SEX", "RACE", "STRATA1")]
trt <- "ARM A"
ref <- "ARM C"
subgroups_levels <- list(RACE = list("Total" = "Race",
                                     "AMERICAN INDIAN OR ALASKA NATIVE" = "American",
                                     "WHITE" = "White",
                                     "ASIAN" = "Asian",
                                     "African" = "BLACK OR AFRICAN AMERICAN"),
                         STRATA1 = list("Total" = "Strata",
                                        "A" = "TypeA",
                                        "B" = "TypeB",
                                        "C" = "Typec"),
                         SEX = list("Total" = "Sex",
                                    "M" = "Male",
                                    "F" = "Female",
                                    "U" = "Unknown"))
g_ae_sub(term,
         id,
         arm,
         arm_sl,
         subgroups,
         subgroups_sl,
         subgroups_levels = subgroups_levels)
}
\author{
Liming Li (Lil128) \email{liming.li@roche.com}
}
