% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_butterfly.R
\name{g_butterfly}
\alias{g_butterfly}
\title{Bbutterfly Plot}
\usage{
g_butterfly(category, groups, block_count = "# of patients",
  block_color = NULL, id, facet_rows = NULL, x_label = block_count,
  y_label = "AE Derived Terms", legend_label = "AETOXGR",
  sort_by = "alphabetical", show_legend = TRUE)
}
\arguments{
\item{category}{vector of y values}

\item{groups}{vector of dichotomization values}

\item{block_count}{string - what to count by (ex: # of AEs or # of patients)}

\item{block_color}{vector - color coding of bar segments}

\item{id}{unique subject identifier variable.}

\item{facet_rows}{vector defines what variable is used to split the
plot into rows, default here is \code{NULL}}

\item{x_label}{string of text for x axis label, default is block_count}

\item{y_label}{string of text for y axis label, default is AE Derived Terms}

\item{sort_by}{character string that defines the ordering of the class and term
variables in the output table,
options: "alphabetical" or "count", default here is set to "count"}

\item{show_legend}{boolean of whether color coding legend is included,
default here is \code{FALSE}}
}
\value{
ggplot object
}
\description{
The butterfly plot is often used in Early Development (ED) and shows ...
}
\examples{
library(random.cdisc.data)
library(dplyr)

ADSL <- radam("ADSL", N=10)
AAE <- radam("AAE", N=10)

data <- left_join(AAE, ADSL, by = c("USUBJID", "STUDYID"))


g_butterfly(category = data$AEBODSYS,
            groups = data$SEX,
            block_count = "# of patients",
            block_color = data$AETOXGR,
            id = data$USUBJID,
            facet_rows = data$RACE,
            x_label = "# of patients",
            y_label = "AE Derived Terms",
            legend_label = "AETOXGR",
            sort_by = "count",
            show_legend = TRUE)

}
\author{
Carolyn Zhang
}
