% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_butterfly.R
\name{g_butterfly}
\alias{g_butterfly}
\title{Create a butterfly plot for Early Development Visualization}
\usage{
g_butterfly(category, groups, block_count = "# of patients",
  block_color = NULL, id, facet_rows = NULL, x_label = block_count,
  y_label = "AE Derived Terms", legend_label = "AETOXGR",
  show_legend = TRUE)
}
\arguments{
\item{category}{vector of y values}

\item{groups}{vector of dichotomization values}

\item{block_count}{string - what to count by (ex: # of AE or # of patients)}

\item{block_color}{color coding of bar segments}

\item{id}{unique subject identifier variable.}

\item{facet_rows}{vector defines what variable is used to split the
plot into rows, default here is NULL}

\item{x_label}{string of text for x axis label, default is block_count}

\item{y_label}{string of text for y axis label, default is AE Derived Terms}

\item{show_legend}{boolean of whether color coding legend is included,
default here is FALSE}
}
\value{
ggplot object
}
\description{
Create a butterfly plot for Early Development Visualization
}
\examples{
library(random.cdisc.data)
library(plyr)
library(dplyr)
library(lemon)
library(reshape2)

data <- left_join(radam("AAE", N=10),radam("ADSL", N=10))
#data <- data \%>\% filter(AEBODSYS \%in\% c("Vascular disorders", "Surgical and medical procedures"))

p <- g_butterfly(category = data$AEBODSYS,
            groups = data$SEX,
            block_count = "# of patients",
            block_color = data$AETOXGR,
            id = data$USUBJID,
            #facet_rows = data$RACE,
            x_label = block_count,
            y_label = "AE Derived Terms",
            legend_label = "AETOXGR",
            show_legend = TRUE)
p

}
\author{
Carolyn Zhang
}
