% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_events_term_id.R
\name{g_events_term_id}
\alias{g_events_term_id}
\title{Common Events Plot}
\usage{
g_events_term_id(
  term,
  id,
  arm,
  arm_sl,
  trt = unique(arm)[1],
  ref = unique(arm)[2],
  sort_by = "term",
  term_selected = NULL,
  rate_range = c(0, 1),
  diff_range = c(-1, 1),
  reversed = FALSE,
  conf_level = 0.95,
  cimethod = "wald",
  axis_side = "left",
  color = c("red", "blue"),
  shape = c(16, 17),
  fontsize = 4,
  title = "Most Frequent On-Therapy Adverse Events",
  foot = "Note: CI is calculated with Wald by default"
)
}
\arguments{
\item{term}{event term vector.}

\item{id}{id of the event term vector. Ususally it is USUBJID.}

\item{arm}{Arm vector. Usually it is ACTARM.}

\item{arm_sl}{Subject level of arm vector. Usually it is adsl$ACTARM.}

\item{trt}{treatment arm}

\item{ref}{reference arm}

\item{sort_by}{sort_by type variable. Default set to "term", to sort_by by term. You can also use
"riskdiff", "meanrisk" to sort.}

\item{term_selected}{selected term to draw common ae. filter term with term_selected if it is not NULL.}

\item{rate_range}{Range for overall rate}

\item{diff_range}{Range for rate difference}

\item{reversed}{Whether to sort_by on reversed variable. Default set to FALSE.}

\item{conf_level}{The confidence interval level, default set to 0.95.}

\item{cimethod}{The method used to calculate confidence interval. Defalt is "wald".
Possible choices are methods supported in `DescTools::BinomDiffCI`.}

\item{axis_side}{the side of the axis label, "left" or "right". Default set to "left".}

\item{color}{Color for the plot. Vector of length 2. Color for arms seperately.
Default set to c("red", "blue")}

\item{shape}{Shape for the plot. Vector of length 2. Shape for arms seperately.
Default set to c(16, 17).}

\item{fontsize}{font size for the plot. It is the size used in ggplot2 with default unit "mm", if you
want "points" you will need to devide the point number by "ggplot2:::.pt"}

\item{title}{the title for the plot. Set to NULL to remove the title. Default set to
"Most Frequent On-Therapy Adverse Events Sorted by $\{sort_by\}".
The "$\{sort_by\}" is used to show the sort_by variable in the title.}

\item{foot}{the footnote of the figure.}
}
\value{
grob object
}
\description{
This function plots commonly occurred events by number of unique subjects with events.
It creates basic summary of events and compares event occurrences between treatment and reference
arms, and can be used for events data such as Adverse Events.
}
\details{
there is no equivalent STREAM output
}
\examples{
library(osprey)
library(random.cdisc.data)

term <- as.character(cadae$AEDECOD)
id <- cadae$USUBJID
arm <- as.character(cadae$ACTARMCD)
arm_sl <- as.character(cadsl$ACTARMCD)
trt <- "ARM A"
ref <- "ARM B"
g_events_term_id(term, id, arm, arm_sl, sort_by = "riskdiff")
g_events_term_id(term, id, arm, arm_sl, sort_by = "term", reversed = FALSE)
g_events_term_id(term, id, arm, arm_sl, sort_by = "meanrisk",
reversed = FALSE, axis_side = "right")
g_events_term_id(term, id, arm, arm_sl, conf_level = 0.9, fontsize = 6)
}
