% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_events_term_id.R
\name{g_events_term_id}
\alias{g_events_term_id}
\title{Events by Term Plot}
\usage{
g_events_term_id(
  term,
  id,
  arm,
  arm_N,
  ref = levels(arm)[1],
  trt = levels(arm)[2],
  sort_by = c("term", "riskdiff", "meanrisk"),
  rate_range = c(0, 1),
  diff_range = c(-1, 1),
  reversed = FALSE,
  conf_level = 0.95,
  diff_ci_method = c("wald", "waldcc", "ac", "score", "scorecc", "mn", "mee", "blj",
    "ha", "beal"),
  axis_side = c("left", "right"),
  color = c("blue", "red"),
  shape = c(16, 17),
  fontsize = 4,
  draw = TRUE
)
}
\arguments{
\item{term}{\code{character} or \code{factor} vector, or \code{data.frame} \cr
Represents events information. \code{term} can be a \code{data.frame} produced
by \code{create_flag_vars}, with each column being a \code{logical} event indicator}

\item{id}{\code{vector} contains subject identifier. Length of \code{id} must be the
same as the length or number of rows of \code{terms}. Ususally it is \code{ADAE$USUBJID}.}

\item{arm}{\code{factor} vector that contains arm informatiion in analysis data.
For example, \code{ADAE$ACTARMCD}.}

\item{arm_N}{(\code{numeric} vector)\cr
Contains information of the number of patients in the levels of \code{arm}. This is useful
if there are patients that have no adverse events can be accounted for with this argument.}

\item{ref}{\code{character} indicates the name of the reference arm. Default is the first
level of \code{arm}.}

\item{trt}{\code{character} indicates the name of the treatment arm. Default is the second
level of \code{arm}.}

\item{sort_by}{\code{character} indicates how each \code{term} is sorted in the plot.
Choose from "term" for alphabetic terms, "riskdiff" for risk difference, and "meanrisk"
for mean risk. Default is "term".}

\item{rate_range}{Numeric \code{vector} of length 2. Range for overall rate display}

\item{diff_range}{Numeric \code{vector} of length 2. Range for rate difference display}

\item{reversed}{\code{logical} whether to reverse the sorting by \code{sort_by}.
Default is FALSE.}

\item{conf_level}{\code{numeric} The confidence interval level, default is 0.95.}

\item{diff_ci_method}{\code{character} The method used to calculate confidence interval.
Defalt is "wald". Possible choices are methods supported in \code{\link[DescTools]{BinomDiffCI}}.}

\item{axis_side}{\code{character} the side of the axis label, "left" or "right". Default is "left".}

\item{color}{Color for the plot. \code{vector} of length 2. Color for reference and
treatment arms respectively. Default set to \code{c("blue", "red")}.}

\item{shape}{Shape for the plot. \code{vector} of length 2. Shape for reference and
treatment arms respectively. Default set to \code{c(16, 17)} per
\code{\link[ggplot2]{scale_shape}}.}

\item{fontsize}{\code{numeric} font size for the plot. It is the size used in ggplot2 with
default unit "mm", if you want "points" you will need to devide the point number by
\code{ggplot2:::.pt}.}

\item{draw}{\code{logical} whether to draw the plot.}
}
\value{
grob object
}
\description{
This function plots commonly occurred events by number of unique subjects with events.
It creates basic summary of events and compares event occurrences between comparison
and reference arms, and can be used for events data such as Adverse Events.
}
\details{
there is no equivalent STREAM output
}
\examples{
library(random.cdisc.data)
library(grid)

ADSL <- radsl(cached = TRUE)
ADAE <- radae(cached = TRUE)

# add additional dummy causality flags
ADAE <- ADAE \%>\%
  mutate(AEREL1 = (AEREL == "Y" & ACTARM == "A: Drug X")) \%>\%
  mutate(AEREL2 = (AEREL == "Y" & ACTARM == "B: Placebo")) \%>\%
  rtables::var_relabel(
    AEREL1 = "AE related to A: Drug X",
    AEREL2 = "AE related to B: Placebo"
  )

term <- ADAE$AEDECOD
id <- ADAE$USUBJID
arm <- ADAE$ACTARMCD
arm_N <- table(ADSL$ACTARMCD)
ref <- "ARM A"
trt <- "ARM C"
# Example 1
p1 <- g_events_term_id(
  term,
  id,
  arm,
  arm_N
)
grid.newpage()
grid.draw(p1)

# Example 2
p2 <- g_events_term_id(
  term,
  id,
  arm,
  arm_N,
  trt = trt,
  ref = ref,
  sort_by = "riskdiff",
  diff_ci_method = "ac",
  conf_level = 0.9
)
grid.newpage()
grid.draw(p2)

# Example 3
p3 <- g_events_term_id(
  term,
  id,
  arm,
  arm_N,
  sort_by = "meanrisk",
  axis_side = "right",
  fontsize = 5
)
grid.newpage()
grid.draw(p3)

# Example 4
term <- create_flag_vars(ADAE)
g_events_term_id(
  term,
  id,
  arm,
  arm_N,
  fontsize = 3
)

}
\author{
Liming Li (Lil128) \email{liming.li@roche.com}

Molly He (hey59) \email{hey59@gene.com}
}
