% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_heat_bygrade.R
\name{g_heat_bygrade}
\alias{g_heat_bygrade}
\title{Heatmap by Grade}
\usage{
g_heat_bygrade(
  id_var,
  visit_var,
  exp_data,
  anno_data,
  anno_var,
  heat_data,
  heat_color_var,
  heat_color_opt = NULL,
  conmed_data = NULL,
  conmed_var = NULL,
  conmed_opt = NULL,
  xlab = "Visit",
  title = NULL
)
}
\arguments{
\item{id_var}{(`character`)\cr name of the column that contains the unique subject identifier shared by all data
Usually it is \code{"USUBJID"}.}

\item{visit_var}{(`character`)\cr name of the column that contains the analysis visit. Usually it is \code{AVISIT}}

\item{exp_data}{(`dataframe`)\cr exposure data. Usually it is \code{ADEX}.}

\item{anno_data}{(`dataframe`)\cr annotation data contains subject level characteristics. Usually it is \code{ADSL}}

\item{anno_var}{(`vector`) a vector of columns names to include for the annotation}

\item{heat_data}{(`dataframe`)\cr contains the information needed for the text over heatmap
Usually is \code{ADCM}.}

\item{heat_color_var}{(`character`)\cr name of the column that contains the heat grade}

\item{heat_color_opt}{(`vector`)\cr vector defines heat color, default here is \code{NULL}}

\item{conmed_data}{(`dataframe`)\cr concomitant medicine data. Usually it is \code{ADCM}
default is \code{NULL} (no conmed plotted)}

\item{conmed_var}{(`character`)\cr concomitant medicine variable name. Must be included in conmed_data
default is \code{NULL} (no conmed plotted)}

\item{conmed_opt}{(`vector`)\cr default is \code{NULL} (use default color or no conmed plotted)}

\item{xlab}{(`character`)\cr string to be shown as x-axis label, default is \code{"Visit"}}

\item{title}{(`character`)\cr string to be shown as title of the plot,
default is \code{NULL} (no plot title is displayed)}
}
\description{
This function plots heatmap
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 30)
ADEX <- radex(ADSL)
ADAE <- radae(ADSL)
ADCM <- radcm(ADSL)
# add AVISIT in ADAE
ADAE$AVISIT <- sample(unique(ADEX$AVISIT[!is.na(ADEX$AVISIT)]), nrow(ADAE), TRUE)
exp_data <- ADEX \%>\%
  filter(PARCAT1 == "INDIVIDUAL") \%>\%
  group_by(USUBJID) \%>\%
  # create a shorter subject identifier
  mutate(SUBJ = utils::tail(strsplit(USUBJID, "-")[[1]], n = 1)) \%>\%
  ungroup()
anno_data <- ADSL \%>\%
  select(SEX, COUNTRY, USUBJID) \%>\%
  group_by(USUBJID) \%>\%
  mutate(SUBJ = utils::tail(strsplit(USUBJID, "-")[[1]], n = 1)) \%>\%
  ungroup() \%>\%
  select(-USUBJID)

heat_data <- ADAE \%>\%
  select(USUBJID, AVISIT, AETOXGR) \%>\%
  group_by(USUBJID) \%>\%
  mutate(SUBJ = utils::tail(strsplit(USUBJID, "-")[[1]], n = 1)) \%>\%
  ungroup() \%>\%
  select(-USUBJID)
heat_color_opt <- c(
  "No Event" = "gray90",
  "1" = "lightsteelblue1",
  "2" = "steelblue1",
  "3" = "steelblue4",
  "4" = "maroon",
  "5" = "brown4"
  )

# include three conmed levels
ADCM_labs <- rtables::var_labels(ADCM)
# derive ADCM.AVISIT
# this example is randomly assigned, please derive your own study-specific ADCM.AVISIT
ADCM$AVISIT <- sample(unique(exp_data$AVISIT), nrow(ADCM), replace = TRUE)
ADCM <- ADCM \%>\%
  filter(
    CMDECOD == "medname A_1/3" | CMDECOD == "medname A_2/3" | CMDECOD == "medname A_3/3"
    ) \%>\%
  mutate(CMDECOD = factor(CMDECOD, levels = unique(CMDECOD)))
rtables::var_labels(ADCM) <- c(ADCM_labs, "Analysis Visit")
conmed_data <- ADCM \%>\%
  group_by(USUBJID) \%>\%
  mutate(SUBJ = utils::tail(strsplit(USUBJID, "-")[[1]], n = 1))

# example plotting conmed
g_heat_bygrade(
  id_var = "SUBJ",
  visit_var = "AVISIT",
  exp_data,
  anno_data,
  anno_var = names(anno_data),
  heat_data,
  heat_color_var = "AETOXGR",
  heat_color_opt,
  conmed_data,
  conmed_var = "CMDECOD",
  conmed_opt = c("green", "green3", "green4")
  )

# example not plotting conmed
g_heat_bygrade(
  id_var = "SUBJ",
  visit_var = "AVISIT",
  exp_data,
  anno_data,
  anno_var = names(anno_data),
  heat_data,
  heat_color_var = "AETOXGR",
  heat_color_opt
  )
}
\author{
Nina Qi (qit3) \email{qit3@gene.com}

Molly He (hey59) \email{hey59@gene.com}
}
