% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_heat_bygrade.R
\name{g_heat_bygrade}
\alias{g_heat_bygrade}
\title{Heatmap by Grade}
\usage{
g_heat_bygrade(
  id_var,
  exp_data,
  visit_var,
  ongo_var,
  anno_data,
  anno_var,
  heat_data,
  heat_color_var,
  heat_color_opt = NULL,
  conmed_data = NULL,
  conmed_var = NULL,
  conmed_color_opt = NULL,
  xlab = "Visit",
  title = NULL
)
}
\arguments{
\item{id_var}{(\code{character})\cr name of the column that contains the unique subject identifier shared by all data
Usually it is \code{"USUBJID"}.}

\item{exp_data}{(\code{data.frame})\cr exposure data. Usually it is \code{ADEX}.}

\item{visit_var}{(\code{character})\cr name of the column that contains the analysis visit. Usually it is \code{"AVISIT"}}

\item{ongo_var}{(\code{character})\cr name of the column in \code{exp_data} that contains the logical variable
indicating whether the treatment is still ongoing.
Usually it can be derived from \code{EOSSTT}}

\item{anno_data}{(\code{data.frame})\cr annotation data that contains subject level characteristics.
Usually it is \code{ADSL}}

\item{anno_var}{(\code{character}) a vector of columns name(s) to include for the annotation}

\item{heat_data}{(\code{data.frame})\cr data frame that contains the information needed for the text over heatmap
Usually it is \code{ADCM}.}

\item{heat_color_var}{(\code{character})\cr name of the column that contains the heat grade}

\item{heat_color_opt}{optional, (\code{character})\cr a named vector that maps the names to heat colors}

\item{conmed_data}{optional, (\code{data.frame})\cr concomitant medicine data. Usually it is \code{ADCM}
default is \code{NULL} (no \code{conmed} plotted)}

\item{conmed_var}{optional, (\code{character})\cr concomitant medicine variable name. Must be a column name in \code{conmed_data}
when \code{conmed_data} is provided. default is \code{NULL} (no \code{conmed} plotted)}

\item{conmed_color_opt}{optional, (\code{character})\cr vector of color name(s) to \code{conmed_data}}

\item{xlab}{optional, (\code{character})\cr string to be shown as x-axis label, default is \code{"Visit"}}

\item{title}{(\code{character})\cr string to be shown as title of the plot.
default is \code{NULL} (no plot title is displayed)}
}
\description{
This function plots heatmap
}
\examples{
library(scda)
library(dplyr)

cached_data <- synthetic_cdisc_data("latest")
ADSL <- cached_data$adsl \%>\% slice(1:30)
ADEX <- cached_data$adex \%>\% filter(USUBJID \%in\% ADSL$USUBJID)
ADAE <- cached_data$adae \%>\% filter(USUBJID \%in\% ADSL$USUBJID)
ADCM <- cached_data$adcm \%>\% filter(USUBJID \%in\% ADSL$USUBJID)
# function to derive AVISIT from ADEX
add_visit <- function(data_need_visit) {
  visit_dates <- ADEX \%>\%
    filter(PARAMCD == "DOSE") \%>\%
    distinct(USUBJID, AVISIT, ASTDTM) \%>\%
    group_by(USUBJID) \%>\%
    arrange(ASTDTM) \%>\%
    mutate(next_vis = lead(ASTDTM), is_last = ifelse(is.na(next_vis), TRUE, FALSE)) \%>\%
    rename(this_vis = ASTDTM)
  data_visit <- data_need_visit \%>\%
    select(USUBJID, ASTDTM) \%>\%
    left_join(visit_dates, by = "USUBJID") \%>\%
    filter(ASTDTM > this_vis & (ASTDTM < next_vis | is_last == TRUE)) \%>\%
    left_join(data_need_visit)
  return(data_visit)
}
# add AVISIT in ADAE and ADCM
ADAE <- add_visit(ADAE)
ADCM <- add_visit(ADCM)
exp_data <- ADEX \%>\%
  filter(PARCAT1 == "INDIVIDUAL") \%>\%
  group_by(USUBJID) \%>\%
  # create a shorter subject identifier
  mutate(SUBJ = utils::tail(strsplit(USUBJID, "-")[[1]], n = 1)) \%>\%
  mutate(ongo_var = (EOSSTT == "ONGOING")) \%>\%
  ungroup()
anno_data <- ADSL \%>\%
  select(SEX, COUNTRY, USUBJID) \%>\%
  group_by(USUBJID) \%>\%
  mutate(SUBJ = utils::tail(strsplit(USUBJID, "-")[[1]], n = 1)) \%>\%
  ungroup() \%>\%
  select(-USUBJID)
heat_data <- ADAE \%>\%
  select(USUBJID, AVISIT, AETOXGR) \%>\%
  group_by(USUBJID) \%>\%
  mutate(SUBJ = utils::tail(strsplit(USUBJID, "-")[[1]], n = 1)) \%>\%
  ungroup() \%>\%
  select(-USUBJID)
heat_color_opt <- c(
  "No Event" = "gray90",
  "1" = "lightsteelblue1",
  "2" = "steelblue1",
  "3" = "steelblue4",
  "4" = "maroon",
  "5" = "brown4"
)
cmdecod_label <- attr(ADCM[["CMDECOD"]], "label")
ADCM <- ADCM \%>\%
  filter(
    CMDECOD == "medname A_1/3" | CMDECOD == "medname A_2/3" | CMDECOD == "medname A_3/3"
  ) \%>\%
  mutate(CMDECOD = factor(CMDECOD, levels = unique(CMDECOD)))
attr(ADCM[["CMDECOD"]], "label") <- cmdecod_label
conmed_data <- ADCM \%>\%
  group_by(USUBJID) \%>\%
  mutate(SUBJ = utils::tail(strsplit(USUBJID, "-")[[1]], n = 1))
# example plotting conmed
g_heat_bygrade(
  id_var = "SUBJ",
  exp_data,
  visit_var = "AVISIT",
  ongo_var = "ongo_var",
  anno_data,
  anno_var = c("SEX", "COUNTRY"),
  heat_data,
  heat_color_var = "AETOXGR",
  heat_color_opt,
  conmed_data,
  conmed_var = "CMDECOD",
  conmed_color_opt = c("green", "green3", "green4")
)
# example not plotting conmed
g_heat_bygrade(
  id_var = "SUBJ",
  exp_data,
  visit_var = "AVISIT",
  ongo_var = "ongo_var",
  anno_data,
  anno_var = c("SEX", "COUNTRY"),
  heat_data,
  heat_color_var = "AETOXGR",
  heat_color_opt
)

}
\author{
Nina Qi (qit3) \email{qit3@gene.com}

Molly He (hey59) \email{hey59@gene.com}
}
