% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_patient_profile.R
\name{g_patient_profile}
\alias{g_patient_profile}
\title{Patient Profile Plot}
\usage{
g_patient_profile(
  ex = NULL,
  ae = NULL,
  rs = NULL,
  cm = NULL,
  lb = NULL,
  arrow_end_day,
  xlim = c(-28, 365),
  xlab = "Study Day",
  title = "Patient Profile"
)
}
\arguments{
\item{ex}{list may contain
\itemize{
\item \code{data} dataframe for \code{ADEX} domain dataset
\item \code{var} vector to identify each lane of \code{ADEX} domain plot
}}

\item{ae}{list may contain
\itemize{
\item \code{data} dataframe for \code{ADAE} domain dataset
\item \code{var} vector to identify each lane of \code{ADAE} plot
\item \code{line_col} factor vector to specify color for segments of \code{ADAE} plot
\item \code{line_col_legend} string to be displayed as line color legend title of \code{ADAE} plot
\item \code{line_col_opt} aesthetic values to map line color values of \code{ADAE} plot
(named vector to map color values to each name).
If not \code{NULL}, please make sure this contains all possible values for \code{line_col} values,
otherwise color will be assigned by \code{ggplot} default, please note that \code{NULL} needs to be
specified
}}

\item{rs}{list may contain
\itemize{
\item \code{data} dataframe for \code{ADRS} domain dataset
\item \code{var} vector to identify each lane of \code{ADRS} domain plot
}}

\item{cm}{list may contain
\itemize{
\item \code{data} dataframe for \code{ADCM} domain dataset
\item \code{var} vector to identify each lane of \code{ADCM} domain plot
}}

\item{lb}{list may contain
\itemize{
\item \code{data} dataframe for \code{ADLB} domain dataset
\item \code{var} vector to identify each lane of \code{ADLB} domain plot
}}

\item{arrow_end_day}{numeric value indicates the end of arrow when arrows are requested}

\item{xlim}{numeric vector for x-axis limit that will be shared by all domain plots, default is
\code{xlim = c(-28, 365)}}

\item{xlab}{string to be shown as x-axis label, default is \code{"Study Day"}}

\item{title}{string to be shown as title of the plot, default is \code{"Patient Profile"}}
}
\value{
plot object
}
\description{
Patient profile plot provides detailed information for a specific subject participating in the study.
The plot includes relevant data for one subject that can help correlate adverse events, response,
concomitant medications, exposure, and laboratory. The plotting of patient profile is modularized, with
each domain plot generated by function \code{\link{patient_domain_profile}}. This \code{\link{g_patient_profile}}
function assembles all requested domain plots into one patient profile.
\code{ADSL}, \code{ADEX}, \code{ADAE}, \code{ADRS}, \code{ADCM} and \code{ADLB} data must be provided.
The plot output will not include domains with data unspecified
}
\examples{
library(dplyr)
library(nestcolor)
# ADSL
ADSL <- rADSL \%>\%
  filter(USUBJID == rADSL$USUBJID[1]) \%>\%
  mutate(
    TRTSDT = as.Date(TRTSDTM),
    max_date = max(as.Date(LSTALVDT), as.Date(DTHDT), na.rm = TRUE),
    max_day = as.numeric(as.Date(max_date) - as.Date(TRTSDT)) + 1
  ) \%>\%
  select(USUBJID, STUDYID, TRTSDT, max_day)


# ADEX
ADEX <- rADEX \%>\%
  select(USUBJID, STUDYID, ASTDTM, PARCAT2, AVAL, AVALU, PARAMCD)
ADEX <- left_join(ADSL, ADEX, by = c("USUBJID", "STUDYID"))

ADEX <- ADEX \%>\%
  filter(PARAMCD == "DOSE") \%>\%
  arrange(PARCAT2, PARAMCD) \%>\%
  mutate(diff = c(0, diff(AVAL, lag = 1))) \%>\%
  mutate(Modification = case_when(
    diff < 0 ~ "Decrease",
    diff > 0 ~ "Increase",
    diff == 0 ~ "None"
  )) \%>\%
  mutate(ASTDT_dur = as.numeric(
    as.Date(substr(as.character(ASTDTM), 1, 10)) -
      as.Date(TRTSDT) + 1
  ))

# ADAE
ADAE <- rADAE \%>\%
  select(USUBJID, STUDYID, AESOC, AEDECOD, AESER, AETOXGR, AEREL, ASTDY, AENDY)
ADAE <- left_join(ADSL, ADAE, by = c("USUBJID", "STUDYID"))

# ADRS
ADRS <- rADRS \%>\%
  select(USUBJID, STUDYID, PARAMCD, PARAM, AVALC, AVAL, ADY, ADTM)
ADRS <- left_join(ADSL, ADRS, by = c("USUBJID", "STUDYID"))

# ADCM
ADCM <- rADCM \%>\%
  select(USUBJID, STUDYID, ASTDTM, AENDTM, CMDECOD, ASTDY, AENDY)
ADCM <- left_join(ADSL, ADCM, by = c("USUBJID", "STUDYID"))

# ADLB
ADLB <- rADLB \%>\%
  select(
    USUBJID, STUDYID, LBSEQ, PARAMCD, BASETYPE, ADTM,
    ADY, ATPTN, AVISITN, LBTESTCD, ANRIND
  )
ADLB <- left_join(ADSL, ADLB, by = c("USUBJID", "STUDYID"))

ADLB <- ADLB \%>\%
  group_by(USUBJID) \%>\%
  mutate(ANRIND = factor(ANRIND, levels = c("LOW", "NORMAL", "HIGH")))

# Example Patient Profile plot 5 domains
g_patient_profile(
  ex = list(
    data = ADEX,
    var = ADEX$PARCAT2
  ),
  ae = list(
    data = ADAE,
    var = ADAE$AEDECOD,
    line_col = factor(ADAE$AESER),
    line_col_legend = "Serious",
    line_col_opt = c("Y" = "red", "N" = "blue")
  ),
  rs = list(
    data = ADRS,
    var = ADRS$PARAMCD
  ),
  cm = list(
    data = ADCM,
    var = ADCM$CMDECOD
  ),
  lb = list(
    data = ADLB,
    var = ADLB$LBTESTCD
  ),
  arrow_end_day = ADSL$max_day,
  xlim = c(-28, ADSL$max_day),
  xlab = "Study Day",
  title = paste("Patient Profile: ", ADSL$USUBJID)
)

# Example Patient Profile plot without ADCM and ADLB
g_patient_profile(
  ex = list(
    data = ADEX,
    var = ADEX$PARCAT2
  ),
  ae = list(
    data = ADAE,
    var = ADAE$AEDECOD,
    line_col = factor(ADAE$AESER),
    line_col_legend = "Serious",
    line_col_opt = c("Y" = "red", "N" = "blue")
  ),
  rs = list(
    data = ADRS,
    var = ADRS$PARAMCD
  ),
  arrow_end_day = ADSL$max_day,
  xlim = c(-28, ADSL$max_day),
  xlab = "Study Day",
  title = paste("Patient Profile: ", ADSL$USUBJID)
)
}
\seealso{
\code{\link{patient_domain_profile}}
}
\author{
Xuefeng Hou (houx14) \email{houx14@gene.com}

Molly He (hey59) \email{hey59@gene.com}

Ting Qi (qit3) \email{qit3@gene.com}
}
