% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_spiderplot.R
\name{g_spiderplot}
\alias{g_spiderplot}
\title{Spider Plot}
\usage{
g_spiderplot(marker_x, marker_y, line_colby = NULL, marker_shape = NULL,
  marker_shape_opt = NULL, marker_size = 3, datalabel_txt = NULL,
  facet_rows = NULL, facet_columns = NULL, vref_line = NULL,
  href_line = NULL, x_label = "Time (Days)",
  y_label = "Change (\%) from Baseline", show_legend = FALSE, draw = TRUE,
  newpage = TRUE)
}
\arguments{
\item{marker_x}{dataframe with 2 columns,
column 1 is the vector of x values (must be in sorted order) and
column 2 is the vector to group the points together (default
should be USUBJID)}

\item{marker_y}{vector of y values}

\item{line_colby}{vector defines by what variable plot is color coded,
default here is \code{NULL}}

\item{marker_shape}{vector defines by what variable points are shape coded,
, default here is \code{NULL}}

\item{marker_shape_opt}{vector defines marker shape code, default here is \code{NULL}}

\item{marker_size}{size of markers in plot, default here is \code{NULL}}

\item{datalabel_txt}{list defines text (at last time point) and
flag for an arrow annotation
(per defined variable) - elements must be labeled txt_ann/mrkr_all/mrkr_ann
txt_ann - text annotation next to final data point (for text annotation)
mrkr_all - vector of ID's (for annotation marker)
mrkr_ann - vector of ID's (subset of mrkr_all) where arrow is desired to
indicate any study interim points}

\item{facet_rows}{dataframe defines what variable is used to split the
plot into rows, default here is \code{NULL}}

\item{facet_columns}{dataframe defines what variable is used to split the
plot into columns, default here is \code{NULL}}

\item{vref_line}{value defines vertical line overlay
(can be a vector), default here is \code{NULL}}

\item{href_line}{value defines horizontal line overlay
(can be a vector), default here is NULL}

\item{x_label}{string of text for x axis label, default is time}

\item{y_label}{string of text for y axis label, default is % change}

\item{show_legend}{boolean of whether marker legend is included,
default here is \code{FALSE}}
}
\value{
ggplot object
}
\description{
Spider plot is often used in Early Development (ED) and displays individual
patient plot of an endpoint over time by group.
}
\details{
there is no equivalent STREAM output
}
\examples{
# simple example
library(dplyr)

data("rADSL")
data("rADTR")
ADTR <- rADTR \%>\% select(STUDYID, USUBJID, ADY, PCHG, PARAMCD)
ADSL <- rADSL \%>\% select(STUDYID, USUBJID, RACE, SEX, ARM)
ANL <- left_join(ADTR, ADSL, by = c("STUDYID", "USUBJID"))
ANL <- ANL \%>\% filter(PARAMCD == "SLDINV") \%>\% filter(RACE \%in\% c("WHITE", "ASIAN")) \%>\% group_by(USUBJID) \%>\% arrange(ADY)
ANL <- na.omit(ANL)
ANL$USUBJID <- substr(ANL$USUBJID, 14, 18)


p <- g_spiderplot(marker_x = data.frame(day = ANL$ADY, groupby = ANL$USUBJID),
             marker_y = ANL$PCHG,
             line_colby = ANL$USUBJID,
             marker_shape = ANL$RACE,
             marker_shape_opt = c("ASIAN" = 1, "NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER" = 2,
                                  "BLACK OR AFRICAN AMERICAN" = 3, "WHITE" = 4,
                                  "AMERICAN INDIAN OR ALASKA NATIVE" = 5, "UNKNOWN" = 6),
             marker_size = 5,
             datalabel_txt = list(txt_ann = ANL$USUBJID),
             #datalabel_txt = list(txt_ann = ANL$USUBJID, mrkr_all = ANL$USUBJID, mrkr_ann = c("id-1", "id-4", "id-7")),
             #datalabel_txt = list(mrkr_all = ANL$USUBJID, mrkr_ann = c("id-2", "id-4", "id-7")),
             facet_rows = data.frame(sex = ANL$SEX),
             facet_columns = data.frame(arm = ANL$ARM),
             vref_line = c(10, 37),
             href_line = -0.3,
             x_label = "Time (Days)",
             y_label = "Change (\%) from Baseline",
             show_legend = TRUE)
p

\dontrun{
#test discrete x-axis points
dat2 <- dat \%>\% arrange(TUDY) \%>\% mutate(day = as.character(TUDY)) \%>\% as.data.frame()
}
}
\author{
Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}
}
