% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_spiderplot.R
\name{g_spiderplot}
\alias{g_spiderplot}
\title{Create a spiderplot for Early Development Visualization}
\usage{
g_spiderplot(x_label = "Time (Days)", y_label = "Change (\%) from Baseline",
  marker_x, marker_y, marker_color = NULL, marker_color_opt = NULL,
  marker_shape = NULL, marker_shape_opt = NULL, marker_size = 6,
  line_color_colby = NULL, datalabel_txt = NULL, facet_rows = NULL,
  facet_columns = NULL, vref_line = NULL, href_line = NULL,
  show_legend = FALSE, draw = TRUE, newpage = TRUE)
}
\arguments{
\item{x_label}{string of text for x axis label}

\item{y_label}{string of text for y axis label}

\item{marker_x}{vector of x values}

\item{marker_y}{vector of y values}

\item{marker_color}{vector defines by what variable points are color coded,
, default here is NULL}

\item{marker_color_opt}{vector defines marker color code, default here is NULL}

\item{marker_shape}{vector defines by what variable points are shape coded,
, default here is NULL}

\item{marker_shape_opt}{vector defines marker shape code, default here is NULL}

\item{marker_size}{size of markers in plot}

\item{line_color_colby}{vector defines by what variable plot is color coded}

\item{datalabel_txt}{list defines text (at last time point) and flag for discontinued study
(per defined variable) - elements must be labeled one/two/three
one - text annotation next to final data point
two - vector of ID's
three - vector of ID's (subset of two) where arrow is desired to indicate discontinued study}

\item{facet_rows}{vector defines what variable is used to split the
plot into rows, default here is NULL}

\item{facet_columns}{vector defines what variable is used to split the
plot into columns, default here is NULL}

\item{vref_line}{value defines vertical line overlay
(can be a vector), default here is NULL}

\item{href_line}{value defines horizontal line overlay
(can be a vector), default here is NULL}

\item{show_legend}{boolean of whether marker legend is included,
default here is FALSE}
}
\value{
ggplot object
}
\description{
Create a spiderplot for Early Development Visualization
}
\examples{
library(random.cdisc.data)
library(plyr)
library(dplyr)

atr <- left_join(radam("ATR", N=10),radam("ADSL", N=10))
dat <- atr \%>\% filter(PARAMCD == "SUMTGLES")

cmp <- max(dat["TUDY"])
lbl <- apply(dat, 1, function(dat){if(dat["TUDY"] == cmp){paste(dat["USUBJID"], "test-length")}else{""}})
colors <- c("black", "red", "blue", "green", "yellow", "brown")
shapes <- c(0, 1, 2, 3, 4, 5, 6)
map_marker_color <- mapvalues(dat$RACE, from = levels(dat$RACE), to = colors[1:nlevels(dat$RACE)])
map_marker_shape <- mapvalues(dat$RACE, from = levels(dat$RACE), to = shapes[1:nlevels(dat$RACE)])
g_spiderplot(x_label = "Time (Days)",
             y_label = "Change (\%) from Baseline",
             marker_x = dat$TUDY,
             marker_y = dat$PCHG,
             marker_color = dat$USUBJID,
             #marker_color_opt = map_marker_color,
             marker_shape = dat$RACE,
             #marker_shape_opt = map_marker_shape,
             marker_size = 5,
             line_color_colby = dat$USUBJID,
             datalabel_txt = list(one = dat$USUBJID, two = dat$USUBJID, three = c("id-2", "id-4", "id-7")),
             facet_rows = dat$SEX,
             facet_columns = dat$ARM,
             vref_line = c(10, 37),
             href_line = -0.3,
             show_legend = FALSE)

}
\author{
Carolyn Zhang
}
