% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_spiderplot.R
\name{g_spiderplot}
\alias{g_spiderplot}
\title{Create a spiderplot for Early Development Visualization}
\usage{
g_spiderplot(x_label, y_label, marker_x, marker_y, marker_color, marker_shape,
  line_color, datalabel_txt, facet_rows = NULL, facet_columns = NULL,
  vref_line = NULL, href_line = NULL)
}
\arguments{
\item{x_label}{string of text for x axis label}

\item{y_label}{string of text for y axis label}

\item{marker_x}{vector of x values}

\item{marker_y}{vector of y values}

\item{marker_color}{vector defines by what variable points are color coded}

\item{marker_shape}{vector defines by what variable points are shape coded}

\item{line_color}{vector defines by what variable plot is color coded}

\item{datalabel_txt}{vector defines text at last time point for each line}

\item{facet_rows}{vector defines what variable is used to split the
plot into rows, default here is NULL}

\item{facet_columns}{vector defines what variable is used to split the
plot into columns, default here is NULL}

\item{vref_line}{value defines vertical line overlay
(can be a vector), default here is NULL}

\item{href_line}{value defines horizontal line overlay
(can be a vector), default here is NULL}
}
\value{
ggplot object
}
\description{
Create a spiderplot for Early Development Visualization
}
\examples{
library(random.cdisc.data)
library(dplyr)

atr <- left_join(radam("ATR", N=10),radam("ADSL", N=10))
dat <- atr \%>\% filter(PARAMCD == "SUMTGLES")

cmp <- max(dat["TUDY"])
temp <- apply(dat, 1, function(dat){if(dat["TUDY"] == cmp){dat["USUBJID"]}else{""}})
g_spiderplot(x_label = "Time (Days)",
             y_label = "Change (\%) from Baseline",
             marker_x = dat$TUDY,
             marker_y = dat$PCHG,
             marker_color = dat$RACE,
             marker_shape = dat$RACE,
             line_color = dat$USUBJID,
             datalabel_txt = temp,
             facet_rows = dat$SEX,
             facet_columns = dat$ARM,
             vref_line = c(10, 37),
             href_line = -0.3)



}
\author{
Carolyn Zhang
}
