% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_swimlane.R, R/test.R
\name{g_swimlane}
\alias{g_swimlane}
\alias{g_swimlane}
\title{Swimlane Plot
Swimlane bar plot for each id with markers}
\usage{
g_swimlane(bar_by, bar_length, sort_by, col_by, marker_by, marker_shape,
  marker_pos, anno_txt, ytick_at, yref_line, xlab, ylab, title)

g_swimlane(bar_by, bar_length, sort_by, col_by, marker_by, marker_shape,
  marker_pos, anno_txt, ytick_at, yref_line, xlab, ylab, title)
}
\arguments{
\item{bar_by}{variable to identify each bar. Default is \code{ID}.}

\item{bar_length}{length variable for each bar}

\item{sort_by}{variable to sort bars}

\item{col_by}{variable to color bars}

\item{marker_by}{variable to order markers}

\item{marker_shape}{variable to specify shape and color for markers}

\item{marker_pos}{variable to specify position for markers}

\item{anno_txt}{description in the left table}

\item{ytick_at}{break interval of y-axis. It takes a numeric vector or \code{NULL}}

\item{yref_line}{numeric vector to plot reference lines}

\item{xlab}{x label}

\item{ylab}{y label}

\item{title}{title}

\item{bar_by}{variable to identify each bar. Default is \code{ID}.}

\item{bar_length}{length variable for each bar}

\item{sort_by}{variable to sort bars}

\item{col_by}{variable to color bars}

\item{marker_by}{variable to order markers}

\item{marker_shape}{variable to specify shape and color for markers}

\item{marker_pos}{variable to specify position for markers}

\item{anno_txt}{description in the left table}

\item{ytick_at}{break interval of y-axis. It takes a numeric vector or \code{NULL}}

\item{yref_line}{numeric vector to plot reference lines}

\item{xlab}{x label}

\item{ylab}{y label}

\item{title}{title}
}
\value{
plot object

plot object
}
\description{
Swimlane Plot
Swimlane bar plot for each id with markers

Swimlane Plot
Swimlane bar plot for each id with markers
}
\examples{

library(random.cdisc.data)
ASL <- radam("ASL", N=50, start_with = list(TRTDUR = rexp(nrow(ASL), 1/100)))
ARS <- radam("ARS", ADSL = ASL,
start_with = list(
      ADY = rexp(nrow(ASL), 1/80)
   )) \%>\% filter(PARAMCD == "OVRINV")
ANL <- ASL \%>\% left_join(ARS, by = c("STUDYID", "USUBJID")) \%>\%
# pre-process bar length breaks for treatment duration for longitudinal records
group_by(USUBJID) \%>\%
mutate(n = n(), bar_len = TRTDUR/n)
anno_txt <- ANL[, c("ARMCD", "SEX", "RACE")]

g_swimlane(bar_by = ANL$USUBJID,
bar_length = ANL$bar_len ,
sort_by = ANL$ARM ,
col_by = ANL$ARM ,
marker_by = ANL$STUDYID,
marker_shape = ANL$AVALC,
marker_pos = ANL$ADY ,
anno_txt = anno_txt,
yref_line = c(100, 200),
ytick_at = waiver(),
xlab = "Patient ID",
ylab = "Time from First Treatment (Day)",
title = "Swimlane Plot")

\\donotrun{
library(dplyr)

atx <- read.bce("/opt/BIOSTAT/qa/s30103j/libraries/atx.sas7bdat")
asl <- read.bce("/opt/BIOSTAT/qa/s30103j/libraries/asl.sas7bdat")
xars <- read.bce("/opt/BIOSTAT/qa/s30103j/libraries/xars.sas7bdat")
# phase 1A data
# pre-process ASl and ARS to fit in ANL
ATX <- atx \%>\% filter(grepl("1a", APERIDC2))
ASL <- ATX \%>\% select(USUBJID) \%>\%
left_join(asl, by = "USUBJID") \%>\%
filter(SAFFL == "Y")

ARS <- ASL \%>\% select(USUBJID) \%>\%
left_join(xars \%>\% filter(grepl("1a", APERIDC2)), "USUBJID") \%>\%
filter(PARAMCD == "OVRINV") \%>\%
select(USUBJID, ADY, AVALC) \%>\%
mutate(domain = "Response")

ADS <- ASL \%>\%
filter(DISCSTUD == "Y" | !is.na(STDSSDT)) \%>\%
select(USUBJID, STDDRS, STDSDY) \%>\%
rename(ADY = STDSDY, AVALC = STDDRS) \%>\%
mutate(domain = "Discon")

# combine with ASL to generate a length for each record
ANL <- ASL \%>\% select(USUBJID, ARMCD, SEX, CADX, TRTDUR) \%>\%
inner_join(rbind(ARS, ADS), "USUBJID") \%>\%
# pre-process bar length breaks for treatment duration for longitudinal records
group_by(USUBJID) \%>\%
mutate(n = n(), bar_len = TRTDUR/n)
anno_txt_var <- c("ARMCD", "SEX", "CADX")
anno_txt <- ANL[, anno_txt_var]

g_swimlane(bar_by = ANL$USUBJID,
bar_length = ANL$bar_len ,
sort_by = ANL$ARMCD ,
col_by = ANL$ARMCD ,
marker_by = ANL$domain,
marker_shape = ANL$AVALC,
marker_pos = ANL$ADY ,
anno_txt = NULL,
yref_line = c(100, 200, 300),
ytick_at = waiver(),
xlab = "Patient ID",
ylab = "Time from First Treatment (Day)",
title = "Swimlane Plot")

}


library(random.cdisc.data)
ASL <- radam("ASL")
ARS <- radam("ARS", ADSL = ASL)

g_swimlane(bar_by = ARS$USUBJID,
bar_length = ARS$AVAL ,
sort_by = ARS$STUDYID ,
col_by = ARS$STUDYID ,
marker_by = ARS$STUDYID,
marker_shape = ARS$AVALC,
marker_pos = ARS$AVAL ,
anno_txt = NULL,
yref_line = c(5, 10),
ytick_at = waiver(),
xlab = "Patient ID",
ylab = "Time from First Treatment (Day)",
title = "Swimlane Plot")

\\donotrun{
atx <- read.bce("/opt/BIOSTAT/qa/s30103j/libraries/atx.sas7bdat")
asl <- read.bce("/opt/BIOSTAT/qa/s30103j/libraries/asl.sas7bdat")
xars <- read.bce("/opt/BIOSTAT/qa/s30103j/libraries/xars.sas7bdat")
# phase 1A data
# pre-process ASl and ARS to fit in ANL
ATX <- atx \%>\% filter(grepl("1a", APERIDC2))
ASL <- ATX \%>\% select(USUBJID) \%>\%
left_join(asl, by = "USUBJID") \%>\%
filter(SAFFL == "Y")

ARS <- ASL \%>\% select(USUBJID) \%>\%
left_join(xars \%>\% filter(grepl("1a", APERIDC2)), "USUBJID") \%>\%
filter(PARAMCD == "OVRINV") \%>\%
select(USUBJID, ADY, AVALC) \%>\%
mutate(domain = "Response")

ADS <- ASL \%>\%
filter(DISCSTUD == "Y" | !is.na(STDSSDT)) \%>\%
select(USUBJID, STDDRS, STDSDY) \%>\%
rename(ADY = STDSDY, AVALC = STDDRS) \%>\%
mutate(domain = "Discon")

# combine with ASL to generate a length for each record
ANL <- ASL \%>\% select(USUBJID, ARMCD, SEX, CADX, TRTDUR) \%>\%
inner_join(rbind(ARS, ADS), "USUBJID") \%>\%
# pre-process bar length breaks for treatment duration for longitudinal records
group_by(USUBJID) \%>\%
mutate(n = n(), bar_len = TRTDUR/n)
anno_txt_var <- c("ARMCD", "SEX", "CADX")
anno_txt <- ANL[, anno_txt_var]

g_swimlane(bar_by = ANL$USUBJID,
bar_length = ANL$bar_len ,
sort_by = ANL$ARMCD ,
col_by = ANL$ARMCD ,
marker_color = ANL2$domain,
marker_shape = ANL2$AVALC,
marker_pos = ANL2$ADY ,
anno_txt = ANL[, anno_txt_var],
yref_line = c(100, 200, 300),
ytick_at = waiver(),
xlab = "Patient ID",
ylab = "Time from First Treatment (Day)",
title = "Swimlane Plot")

}

}
\author{
qit3

qit3
}
