% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderplot.R
\name{spiderplot}
\alias{spiderplot}
\title{Create a spiderplot}
\usage{
spiderplot(x, y, group, color)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{col}{color of points}
}
\value{
ggplot object
}
\description{
Descr of this plot
}
\examples{
library(random.cdisc.data)


ATE <- radam("ATE", N=10 ,start_with = list(CHANGE = c(0, 80, 85, 90), STUDYID = c("A","B")))

with(ATE, spiderplot(AVAL, CHANGE, USUBJID, STUDYID))

spiderplot(x = ATE$AVAL, y = ATE$CHANGE, group= ATE$USUBJID,color = ATE$STUDYID)


}
\author{
Mika
}
