% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_spiderplot_simple.R
\name{spiderplot_simple}
\alias{spiderplot_simple}
\title{Simple spider plot}
\usage{
spiderplot_simple(
  anl,
  byvar = "USUBJID",
  days = "TRTDURD",
  mes_value = "PARAM",
  group_col = "USUBJID",
  baseday = 0
)
}
\arguments{
\item{anl}{The analysis data frame}

\item{byvar}{Analysis dataset}

\item{days}{Variable with time in days}

\item{mes_value}{Variable with measurement}

\item{group_col}{Variable to color the individual lines and id in plot}

\item{baseday}{Numeric Value, points with only smaller values will be cut out}
}
\value{
\code{ggplot} object
}
\description{
Description of this plot
}
\examples{
library(dplyr)
library(nestcolor)

ADSL <- osprey::rADSL[1:15, ]
ADTR <- osprey::rADTR
ANL <- left_join(ADSL, ADTR)

ANL \%>\%
  dplyr::filter(ANL01FL == "Y" & PARAMCD == "SLDINV") \%>\%
  spiderplot_simple(group_col = "SEX", days = "ADY", mes_value = "AVAL")
}
\author{
Mika Maekinen
}
