% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_spiderplot_simple.R
\name{spiderplot_simple}
\alias{spiderplot_simple}
\title{Create a spiderplot_simple}
\usage{
spiderplot_simple(anl, byvar = "USUBJID", days = "TRTDURD",
  mesValue = "PARAM", groupCol = "USUBJID", baseday = 0)
}
\arguments{
\item{anl}{The analysis data frame (e.g. ATE.sas7bdat on BCE)}

\item{byvar}{Analysis dataset}

\item{days}{Variable with time in days}

\item{mesValue}{Variable with measurement}

\item{groupCol}{Variable to color the individual lines and id in plot}

\item{baseday}{Numeric Value, pts with only smaller values will be cut out}
}
\value{
ggplot object
}
\description{
Descr of this plot
}
\examples{
library(random.cdisc.data)

atr <- left_join(rADRS, rADSL)

atr \%>\%
  filter(PARAMCD == "OVRINV") \%>\%
  spiderplot_simple(, groupCol = "SEX", days = "ADY", mesValue = "AVAL")
}
\author{
Mika Maekinen
}
