% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_spiderplot_simple.R
\name{spiderplot_simple}
\alias{spiderplot_simple}
\title{Simple spider plot}
\usage{
spiderplot_simple(anl, byvar = "USUBJID", days = "TRTDURD",
  mes_value = "PARAM", group_col = "USUBJID", baseday = 0)
}
\arguments{
\item{anl}{The analysis data frame (e.g. ATE.sas7bdat on BCE)}

\item{byvar}{Analysis dataset}

\item{days}{Variable with time in days}

\item{mes_value}{Variable with measurement}

\item{group_col}{Variable to color the individual lines and id in plot}

\item{baseday}{Numeric Value, pts with only smaller values will be cut out}
}
\value{
ggplot object
}
\description{
Descr of this plot
}
\examples{
library(random.cdisc.data)
library(dplyr)

ADSL <- rADSL
ADRS <- rADRS
atr <- left_join(ADSL, ADRS)

atr \%>\%
  dplyr::filter(PARAMCD == "OVRINV") \%>\%
  spiderplot_simple(group_col = "SEX", days = "ADY", mes_value = "AVAL")
}
\author{
Mika Maekinen
}
