% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_filter.R
\name{stream_filter}
\alias{stream_filter}
\title{Applies STREAM style filtering to datasets}
\usage{
stream_filter(
  slref = NULL,
  anl = NULL,
  filters,
  suffix,
  slref_keep = NULL,
  usubjid = "USUBJID"
)
}
\arguments{
\item{slref}{The subject level data frame (typically \code{ADSL})}

\item{anl}{The analysis data frame}

\item{filters}{The name of the filters dataset}

\item{suffix}{The suffix to apply in quotes (e.g. \code{"ITT_PFSINV"})}

\item{slref_keep}{Variables to keep from \code{slref} (e.g. \code{c("REGION", "SEX")})}

\item{usubjid}{The unique subject identifier variable in quotes (e.g. \code{"USUBJID"})}
}
\value{
\code{dataframe} object
}
\description{
One of \code{slref} or \code{anl} need to be specified. The conversion from \code{SAS} code in filters dataset may not work in all
cases. In case of failure a sensible error message should be returned.
}
\examples{
ADSL <- rADSL
ADTTE <- rADTTE
filters <- as.data.frame(rbind(
  c(ID = "IT", FLTTARGET = "SLREF", FLTWHERE = "where 1 eq 1"),
  c(ID = "BIO", FLTTARGET = "SLREF", FLTWHERE = "where BMRKR1 ge 4.3"),
  c(ID = "M", FLTTARGET = "SLREF", FLTWHERE = "where SEX eq 'M'"),
  c(ID = "PFS", FLTTARGET = "ANL", FLTWHERE = "where PARAMCD eq 'PFS'"),
  c(ID = "OS", FLTTARGET = "ANL", FLTWHERE = "where PARAMCD eq 'OS'")
))

ANL <- stream_filter(
  slref = ADSL,
  anl = ADTTE,
  suffix = "IT_PFS_BIO",
  filters = filters
)
}
\author{
Iain Bennett
}
