% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae.R
\name{t_ae}
\alias{t_ae}
\title{Adverse Events by System Organ Class and Preferred Term}
\usage{
t_ae(class, term, id, col_by, total = "All Patients")
}
\arguments{
\item{class}{system organ class variable.}

\item{term}{preferred term variable.}

\item{id}{unique subject identifier variable. If a particular subject has no
adverse event then the subject \code{id} should be listed where
\code{class} and \code{term} should be set to missing (i.e. \code{NA}).}

\item{col_by}{group variable that will be used for a column header. \code{col_by}
has to be a factor and can not be missing.}

\item{total}{character string that will be used as a label for a column with
pooled total population, default here is "All Patients", if set to \code{NULL} then
the "All Patients" column is suppressed.}
}
\value{
\code{rtable} object
}
\description{
\code{t_ae} returns adverse events summary table that corresponds to STREAM template AET02
}
\details{
this is an equivalent of the STREAM output \code{\%stream_t_summary(templates = aet02)}
  (\url{http://bioportal.roche.com/stream_doc/2_05/um/report_outputs_aet02.html})
}
\examples{
# Simple example
library(rtables)
library(dplyr)

ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = paste("ARM", LETTERS[rep(c(1, 2), c(3, 7))])
)


ae_lookup <- tribble(
  ~CLASS, ~TERM, ~GRADE,
  "cl A", "trm A_1/2", 1,
  "cl A", "trm A_2/2", 2,
  "cl B", "trm B_1/3", 2,
  "cl B", "trm B_2/3", 3,
  "cl B", "trm B_3/3", 1,
  "cl C", "trm C_1/1", 1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2, 2, 2, 3, 3, 4, 4, 4, 4, 5, 6, 6, 7, 7)]
  ),
  ae_lookup[c(1, 1, 2, 6, 4, 2, 2, 3, 4, 2, 1, 5, 4, 6), ]
)

ANL <- left_join(ASL, AAE, by = "USUBJID")



tbl <- t_ae(
  class = ANL$CLASS,
  term = ANL$TERM,
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  total = "All Patients"
)
tbl

# Simple example 2

ADSL <- rADSL \%>\% select(USUBJID, STUDYID, ARM)
AAE <- rADAE \%>\% select(USUBJID, STUDYID, ARM, AEBODSYS, AEDECOD)
ANL <- left_join(AAE, ADSL, by = c("USUBJID", "STUDYID", "ARM"))


tbl2 <- t_ae(
  class = ANL$AEBODSYS,
  term = ANL$AEDECOD,
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  total = NULL
)
tbl2

}
\author{
Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}
}
