% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae_oview.R
\name{t_ae_oview}
\alias{t_ae_oview}
\title{AE Overview Summary Table}
\usage{
t_ae_oview(id, class, term, flags, display_id = c("fatal", "ser", "serwd",
  "serdsm", "relser", "wd", "dsm", "rel", "relwd", "reldsm", "ctc35"),
  extra_flag = NULL, col_by, total = "All Patients")
}
\arguments{
\item{id}{unique subject identifier variable. If a particular subject has no
adverse event then the subject \code{id} should be listed where
\code{class} and \code{term} should be set to missing (i.e. \code{NA}).}

\item{class}{system organ class variable.}

\item{term}{preferred term variable.}

\item{flags}{dataframe containing all event flags of interest,
default should include dthfl ,dcsreas, aesdth, aeser, aeacn, arel, aerel,
atoxgr to match STREAM output (lowercase column names).
Minimally needs: dthfl and dcsreas}

\item{display_id}{vector of strings for which analyses to display
possible values are: fatal ser serwd serdsm relser wd dsm rel relwd reldsm ctc35}

\item{extra_flag}{dataframe with flags to be calculated (column name here
will be the table's row name)}

\item{col_by}{group variable that will be used for a column header. \code{col_by}
has to be a factor and can not be missing.}

\item{total}{character string that will be used as a label for a column with
pooled total population, default here is "All Patients", if set to \code{NULL} then
the "All Patients" column is suppressed.}
}
\value{
\code{rtable} object
}
\description{
\code{t_ae_oview} returns a summary table of overall AE profile: adverse events, deaths,
and withdrawals by trial treatment. It corresponds to STREAM template AET01.
}
\details{
this is an equivalent of the STREAM output \code{\%stream_t_summary(templates = aet01)}
  (\url{man}{http://bioportal.roche.com/stream_doc/2_05/um/report_outputs_aet01.html})
}
\examples{
library(dplyr)

data("rADAE")
ANL <- rADAE
flag <- data.frame(dthfl = ANL$DTHFL,
                   dcsreas = ANL$DCSREAS,
                   aesdth = ANL$AESDTH,
                   aeser = ANL$AESER,
                   aeacn = ANL$AEACN,
                   arel = ANL$AREL,
                   aerel = ANL$AEREL,
                   aetoxgr = ANL$AETOXGR)

extra <- data.frame(fatal2 = flag$aesdth,
                   ser2 = flag$aeser)
tbl <- t_ae_oview(
   id = ANL$USUBJID,
   class = ANL$AESOC,
   term = ANL$AEDECOD,
   flags = flag,
   display_id = c("fatal", "ser", "serwd", "serdsm", "relser",
                  "wd", "dsm", "rel", "relwd", "reldsm"),
   extra_flag = extra,
   col_by = factor(ANL$ARM),
   total="All Patients"
)

tbl


}
\author{
Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}
}
