% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ds.R
\name{t_ds}
\alias{t_ds}
\title{Patient Disposition Table}
\usage{
t_ds(class, term, sub = NULL, id, col_by, total = "All Patients", ...)
}
\arguments{
\item{term}{reason for discontinuation from study.}

\item{sub}{dataframe of additional subsets of terms}

\item{id}{unique subject identifier variable. If a particular subject has no
adverse event then the subject \code{id} should be listed where
\code{class} and \code{term} should be set to missing (i.e. \code{NA}).}

\item{col_by}{group variable that will be used for a column header. \code{col_by}
has to be a factor and can not be missing.}

\item{total}{character string that will be used as a label for a column with
pooled total population, default here is "All Patients", if set to "NONE" then
the "All Patients" column is suppressed.}
}
\value{
\code{rtable} object
}
\description{
\code{t_ds} returns patient disposition table that corresponds to STREAM template DST01
}
\details{
this is an equivalent of the STREAM output \code{\%stream_t_summary(templates = aet01)}
  (\url{man}{http://bioportal.roche.com/stream_doc/2_05/um/report_outputs_aet01.html})
}
\examples{
# Simple example
library(dplyr)

ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = paste("ARM", LETTERS[rep(c(1,2), c(3,7))])
)


ae_lookup <- tribble(
~CLASS,         ~TERM,   ~GRADE,
"cl A",   "trm A_1/2",        1,
"cl A",   "trm A_2/2",        2,
"cl B",   "trm B_1/3",        2,
"cl B",   "trm B_2/3",        3,
"cl B",   "trm B_3/3",        1,
"cl C",   "trm C_1/1",        1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2,2,2,3,3,4,4,4,4,5,6,6,7,7)]
  ),
  ae_lookup[c(1,1,2,6,4,2,2,3,4,2,1,5,4,6),]
)

ANL <- left_join(ASL, AAE, by = "USUBJID")

tbl <- t_ds(
  class = ANL$CLASS,
  term =  ANL$TERM,
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  total = "All Patients"
)

tbl

tbl2 <- t_ds(
  class = ANL$CLASS,
  term =  ANL$TERM,
  sub = data.frame(GRADE = ANL$GRADE),
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  total = "All Patients"
)

tbl2

# simple example using osprey dummy dataset

data("rADSL")
ANL <- rADSL

tbl3 <- t_ds(
   class = ANL$EOSSTT,
   term = ANL$DCSREAS,
   id = ANL$USUBJID,
   col_by = factor(ANL$ARM),
   total = "All Patients"
)

tbl3


}
\author{
Carolyn Zhang (zhanc107) \email{carolyn.zhang@duke.edu}
}
